% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_constraint_freeway_ramp_junction.R
\name{Analysis_availability_freeway_ramp_junction}
\alias{Analysis_availability_freeway_ramp_junction}
\title{Analysis Constraint of Ramp Freeway Junction}
\usage{
Analysis_availability_freeway_ramp_junction(
  free_speed = NULL,
  N = NULL,
  C = NULL
)
}
\arguments{
\item{free_speed}{*Numeric* Free speed in connecting road(kph)}

\item{N}{*Categorical* The number of one-way lane of connecting road. Choose one from : \code{1}, \code{2}}

\item{C}{*Numeric* Capacity of the connecting road(pcph)}
}
\description{
It follows <Table 4-2> in KHCM(2013). The suggested capacity value indicates whether or not to proceed with the service level analysis procedure of the connection. Used to judge. If at least one of the points where the capacity value presented in the table exceeds the capacity value, the service level analysis procedure is not performed. When the dose values at the points indicated in the presented table are not exceeded, proceed with the analysis procedure. When all values are not exceeded, the analysis procedure is performed.
}
\examples{
Analysis_availability_freeway_ramp_junction(free_speed = 80, N = 2, C = 4000)
Analysis_availability_freeway_ramp_junction(50, 1, 3000)
}
