% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATS_1_i.R
\name{ATS_1_i}
\alias{ATS_1_i}
\title{Average Travel Speed in type_1 2-lane road, i Section(ATS_1_i, kph)}
\usage{
ATS_1_i(FFS = NULL, V_d = NULL, V_o = NULL, f_np_ATS = NULL, f_w_ATS = NULL)
}
\arguments{
\item{FFS}{Free speed(kph)}

\item{V_d}{*Numeric* Traffic in the direction of travel(pcph?)}

\item{V_o}{*Numeric* Counter traffic(pcph?)}

\item{f_np_ATS}{*Numeric* See <Table 7-12>}

\item{f_w_ATS}{*Numeric* See <Table 7-11>}
}
\description{
This function follows <Formula 6-2>
}
