% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATS_2lane_plus1_TD.R
\name{ATS_2lane_plus1_TD}
\alias{ATS_2lane_plus1_TD}
\title{Average travel speed of a 2+1 lane road in Total Section(ATS_2lane_plus1_TD, kph)}
\usage{
ATS_2lane_plus1_TD(
  ATS_2lane_plus1_IN = NULL,
  ATS_2lane_plus1_OUT = NULL,
  L_IN = NULL,
  L_OUT = NULL
)
}
\arguments{
\item{ATS_2lane_plus1_IN}{*Numeric* ATS in the direction of progress}

\item{ATS_2lane_plus1_OUT}{*Numeric* ATS in opposite direction}

\item{L_IN}{*Numeric* Section length for analysis in the direction of progress(m)}

\item{L_OUT}{*Numeric* Section length for analysis in the opposite direction(m)}
}
\description{
This function follows <Formula 7-16>
}
\examples{
ATS_2lane_plus1_TD(70, 55, 0.4, 1.3)
}
