% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_speed.R
\name{avg_speed}
\alias{avg_speed}
\title{Average moving speed divided into 5 sections (acceleration 4km/h)}
\usage{
avg_speed(d_sum = NULL, unit = NULL)
}
\arguments{
\item{d_sum}{d_T + d_WZ + d_IW}

\item{unit}{Choose one from : \code{'kmph'}, \code{'mps'}}
}
\description{
It follows <Table 8-21> in KHCM(2013)
}
\examples{
avg_speed(130, 'mps')
avg_speed(88.33, 'kmph')
}
