% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_speed_arterial.R
\name{avg_speed_arterial}
\alias{avg_speed_arterial}
\title{Average travel speed on arterial roads(avg_speed_arterial, kph)}
\usage{
avg_speed_arterial(L = NULL, t_p_km_arterial = NULL, total_d = NULL)
}
\arguments{
\item{L}{section length(km)}

\item{t_p_km_arterial}{Total traveling time per km of all or part of the arterial road (sec/km). See t_p_km_arterial()}

\item{total_d}{Total approach delay (seconds) at intersections within the scope of analysis for all or part of the arterial road}
}
\description{
It follows <Formula 12-1> in KHCM(2013) p.535. 3600 means Conversion factor for converting speed units of m/sec to km/hour
}
