% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_speed_segment.R
\name{avg_speed_segment}
\alias{avg_speed_segment}
\title{Average travel speed of the entire arterial road with exclusive median bus lanes(avg_speed_segment, kph)}
\usage{
avg_speed_segment(
  L = NULL,
  T_bus = NULL,
  T_others = NULL,
  V_bus = NULL,
  V_others = NULL
)
}
\arguments{
\item{L}{section length(km)}

\item{T_bus}{See T_bus(). Travel time (seconds) of the analysis section of the exclusive median bus lane}

\item{T_others}{See T_others(). Travel time (seconds) of the general lane analysis section}

\item{V_bus}{Peak traffic volume (large/hour) in the analysis section of the exclusive median bus lane}

\item{V_others}{Peak traffic volume of general lane analysis section (large/hour)}
}
\description{
It follows <Formula 12-12> in KHCM(2013) p.552. 3600 means Conversion factor for converting speed units of m/sec to km/hour
}
