% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_multilane_road.R
\name{B_multilane_road}
\alias{B_multilane_road}
\title{Bendiness in Multi-lane Road(B, degree/km)}
\usage{
B_multilane_road(angles = NULL, L = NULL)
}
\arguments{
\item{angles}{*Series* Series of each direct angle. \code{c(theta1, theta2, theta3, ...)} (degree)}

\item{L}{*Numeric* Length of the route(km). Generally, \code{L = 3}}
}
\description{
This function calculates horizontal curve curvature in multi-lane road. It follows <Formula 6-1> in KHCM(2013).
}
\examples{
B_multilane_road(angles = c(30, 20, 30, 40, 50), L = 3)
B_multilane_road(c(10, 8, 12, 14), 5)
}
\keyword{Multilane}
\keyword{bendiness}
\keyword{road}
