% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_capacity_in_construction_section.R
\name{basic_capacity_in_construction}
\alias{basic_capacity_in_construction}
\title{Basic capacity of construction section by design speed(pcphpl)}
\usage{
basic_capacity_in_construction(design_speed = NULL)
}
\arguments{
\item{design_speed}{*Categorical* Choose one from : \code{120}, \code{100}, \code{80} (kph)}
}
\description{
This function decides Basic capacity of construction section by design speed(pcphpl). The contents of <Table 2-7> of KHCM(2013) were reflected.
}
\examples{
basic_capacity_in_construction(80)
basic_capacity_in_construction(design_speed = 120)
}
