% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_b.R
\name{c_b}
\alias{c_b}
\title{Maximum number of buses per hour per bus stopping area (c_b, vph)}
\usage{
c_b(flow_type = NULL, R = NULL, t_c = NULL, t_D = NULL, g_c_ratio = NULL)
}
\arguments{
\item{flow_type}{Choose one from: \code{'continuous'}, \code{'interrupted'}}

\item{R}{Parking surface capacity correction factor. See R()}

\item{t_c}{Erasing time (sec)}

\item{t_D}{Stop time (seconds) (= door opening/closing time + passenger boarding and disembarking time)}
}
\description{
It follows <Formula 13-1>, <Formula 13-2> KHCM(2013), p.598, p.599
}
