% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_Bv.R
\name{c_Bv}
\alias{c_Bv}
\title{Maximum number of vehicles per hour per stop(c_Bv)}
\usage{
c_Bv(
  flow_type = NULL,
  R = NULL,
  t_c = NULL,
  t_D = NULL,
  g_c_ratio = NULL,
  N = NULL
)
}
\arguments{
\item{flow_type}{Choose one from: \code{'continuous'}, \code{'interrupted'}}

\item{R}{Parking surface capacity correction factor. See R()}

\item{t_c}{Erasing time (sec)}

\item{t_D}{Stop time (seconds) (= door opening/closing time + passenger boarding and disembarking time)}

\item{N}{See N_bus_using_efficiency(). Utilization efficiency coefficient according to the number of stopping surfaces (or the length of stopping surfaces)}
}
\description{
It follows <Formula 13-3>, <Formula 13-4> KHCM(2013), p.601
}
