% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_i.R
\name{c_i}
\alias{c_i}
\title{Capacity of the i-th lane group(c_i, vph)}
\usage{
c_i(S_i = NULL, g_i = NULL, C = NULL)
}
\arguments{
\item{S_i}{Saturated traffic flow rate (vph) for the i lane group}

\item{g_i}{Effective green time for i lane group (seconds)}

\item{C}{signal period(seconds)}
}
\description{
It follows <Formula 8-40> in KHCM(2013)
}
