% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_i_diamond_interchange.R
\name{c_i_diamond_interchange}
\alias{c_i_diamond_interchange}
\title{Capacity of the Diamond Interchange(c_i_diamond_interchange, vph)}
\usage{
c_i_diamond_interchange(S_i = NULL, g_i_backtick = NULL, C = NULL)
}
\arguments{
\item{S_i}{Saturated traffic flow rate (vph) for the i lane group}

\item{g_i_backtick}{Effective green time (seconds) for the i lane group reflecting the additional green loss time}

\item{C}{signal period(seconds)}
}
\description{
It follows <Formula 9-9> in KHCM(2013) p.433
}
