% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_i_WZ.R
\name{c_i_WZ}
\alias{c_i_WZ}
\title{Correction of saturated traffic flow rate due to road occupation construction near intersections(S_i_WZ)}
\usage{
c_i_WZ(S_i_WZ = NULL, g_i = NULL, C = NULL)
}
\arguments{
\item{g_i}{Effective green time of lane group i(s)}
}
\description{
It follows <Formula 8-67> in KHCM(2013)
}
