% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_p_x.R
\name{c_p_x}
\alias{c_p_x}
\title{Potential capacity of moving flow x (c_p_x, pcph)}
\usage{
c_p_x(V_c_x = NULL, t_c_x = NULL, t_f_x = NULL)
}
\arguments{
\item{V_c_x}{Conflict flow for flow x. see V_ci()}

\item{t_c_x}{Critical spacing for moving flow x. See t_c_x()}

\item{t_f_x}{Follow-up time for flow x. See t_f_x()}
}
\description{
It follows <Formula 10-1> in KHCM(2013), p.466
}
\examples{
c_p_x(V_c_x = 1200, t_c_x = 4.6, t_f_x = 3.0)
c_p_x(832, 5.4, 2.7)
}
