% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_roundabout.R
\name{c_roundabout}
\alias{c_roundabout}
\title{Roundabout ramp capacity(c_roundabout, pcph)}
\usage{
c_roundabout(lane = NULL, V_c = NULL, f_p = NULL)
}
\arguments{
\item{lane}{roundabout type. 'rotation lane - entry lane'. Choose one from : \code{'1-1'}, \code{'2-2'}, \code{'2-1'}}

\item{V_c}{Conflict traffic volume(pcph)}

\item{f_p}{Crossing Pedestrian Influence Coefficient. See f_p()}
}
\description{
It follows <Formula 11-2> in KHCM(2013) p.496, <Table 11-8> in p.507
}
\examples{
c_roundabout(lane = '1-1', V_c = 382, f_p = 0.8)
}
