% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_SH.R
\name{c_SH}
\alias{c_SH}
\title{Capacity in allotted lanes (c_SH, pcph)}
\usage{
c_SH(V_l = NULL, V_t = NULL, V_r = NULL, c_ml = NULL, c_mt = NULL, c_mr = NULL)
}
\arguments{
\item{V_l}{Traffic volume or flow rate (pcph) distributed to left-turn lanes}

\item{V_t}{Traffic volume or traffic flow rate (pcph) distributed to straight lanes}

\item{V_r}{Traffic volume or traffic flow rate (pcph) allocated to the right turn lane}

\item{c_ml}{Left-turn travel capacity allocated to the lane (pcph)}

\item{c_mt}{Forwarding capacity distributed to lanes (pcph)}

\item{c_mr}{Right-turn travel capacity allocated to lanes (pcph)}
}
\description{
This function follows <Formula 10-4> in KHCM(2013) p.469
}
