% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_wz.R
\name{c_wz}
\alias{c_wz}
\title{Construction section capacity(c_wz, vph)}
\usage{
c_wz(design_speed = NULL, N = NULL, f_w = NULL, f_hv = NULL)
}
\arguments{
\item{design_speed}{*Categorical* Choose one from : \code{120}, \code{100}, \code{80} (kph)}

\item{N}{*Numeric* The number of freeway lane(one-way lane). It means (# of normal one-way lane) - (# of lanes closed due to construction)}

\item{f_w}{*Numeric* The One Side Lane Width and Lateral Clearance Factor(f_w). It could be calculated by f_w()}

\item{f_hv}{*Numeric* Heavy Vehicle Factors(f_hv). See f_hv()}
}
\description{
This function calculates the construction section capacity(vph). It follows <Formula 2-10> in KHCM(2013)
}
\examples{
c_wz(design_speed = 100, N = 3, f_w = 0.9, f_hv = 0.6)
c_wz(80, 4, 0.9, 0.85)
}
