% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity.R
\name{capacity}
\alias{capacity}
\title{Capacity(c, vph)}
\usage{
capacity(
  design_speed = NULL,
  N = NULL,
  f_w = NULL,
  f_hv = NULL,
  weather = "sunny",
  precipitation = NULL,
  day_night = "day"
)
}
\arguments{
\item{design_speed}{*Categorical*}

\item{N}{*Numeric* The number of freeway lane(one-way lane). It must be 2 or more.}

\item{f_w}{*Numeric* The One Side Lane Width and Lateral Clearance Factor(f_w). It could be calculated by f_w()}

\item{f_hv}{*Numeric* Heavy Vehicle Factors(f_hv). See f_hv()}

\item{weather}{*Categorical* Choose one from : \code{'sunny'}, \code{'rainy'}, \code{'snowy'}}

\item{day_night}{*Categorical* Choose one from : \code{'day'}, \code{'night'}}
}
\description{
This function calculates the capacity(vph) for a given road and traffic condition.
}
\examples{
capacity(design_speed = 100, c_j = 1500, N = 4, f_w = 0.79, f_hv = 0.62)
capacity(design_speed = 120, c_j = 1000, N = 3, f_w = 0.8, f_hv = 0.5, weather = 'sunny', day_night = 'day')
capacity(design_speed = 100, c_j = 900, N = 4, f_w = 0.82, f_hv = 0.63, weather = 'rainy', precipitation = 8.2, day_night = 'night')
}
