% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_arterial_road.R
\name{capacity_arterial_road}
\alias{capacity_arterial_road}
\title{The Capacity of Arterial Road(vph)}
\usage{
capacity_arterial_road(N = NULL, g_c_ratio = NULL)
}
\arguments{
\item{N}{Number of lanes}

\item{g_c_ratio}{Green time signal cycle ratio}
}
\description{
This function follows <Formula 12-7> in KHCM(2013) p.540
}
\examples{
capacity_arterial_road(3, 0.2)
}
