% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_freeway_weaving.R
\name{capacity_freeway_weaving}
\alias{capacity_freeway_weaving}
\title{The Capacity of the Weaving Section Between the Main Line and the Connecting Road(pcph)}
\usage{
capacity_freeway_weaving(
  design_speed = NULL,
  V = NULL,
  V_w = NULL,
  L = NULL,
  N = NULL
)
}
\arguments{
\item{design_speed}{\emph{Categorical} Choose one from : \code{120}, \code{100}, \code{80} (kph)}

\item{V}{\emph{Numeric} Total traffic volume (pcph) of the weaving section}

\item{V_w}{\emph{Numeric} Weaving traffic volume(pcph)}

\item{L}{\emph{Numeric} Length of the weaving section(m)}

\item{N}{\emph{Categorical} Total number of lanes in the weaving section. Choose one from : \code{3}, \code{4}, \code{5}}
}
\description{
This function decides The capacity of the weaving section between the main line and the connecting road(pcph). It follows <Table 3-2> in KHCM(2013).
}
\keyword{capacity}
\keyword{freeway}
\keyword{line}
\keyword{link}
\keyword{main}
\keyword{section}
\keyword{weaving}
