% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity_multilane_road.R
\name{capacity_multilane_road}
\alias{capacity_multilane_road}
\title{The Capacity of Multi-lane Road(pcph)}
\usage{
capacity_multilane_road(
  type = NULL,
  design_speed = NULL,
  N = NULL,
  S = NULL,
  g_c_ratio = NULL
)
}
\arguments{
\item{type}{*Categorical* Type of traffic flow in multi-lane road. Choose one from : \code{'interrupted_flow'}, \code{'uninterrupted_flow'}}

\item{design_speed}{*Categorical* If \code{type == 'uninterrupted_flow'}, design_speed must be used.}

\item{N}{*Numeric* Number of lanes going straight at the intersection.}

\item{S}{*Numeric* Saturation flow(pcphpl)}

\item{g_c_ratio}{*Numeric* Average green time ratio}
}
\description{
This function follows <Formula 6-2>
}
\examples{
capacity_multilane_road(type = 'uninterrupted_flow', design_speed = 80)
capacity_multilane_road(type = 'interrupted_flow', N = 3, S = 2000, g_c_ratio = 0.2)
}
\keyword{capacity}
\keyword{freeway}
\keyword{line}
\keyword{link}
\keyword{main}
\keyword{section}
\keyword{weaving}
