% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRF.R
\name{CRF}
\alias{CRF}
\title{Capacity reduction rate index by construction type(CRF)}
\usage{
CRF(construction_type = NULL, lane_reduce = NULL, width_reduce = NULL)
}
\arguments{
\item{construction_type}{Choose one from: \code{'lane_reduce'}, \code{'width_reduce'}, \code{'lane_moved'}, \code{'court_board'}}

\item{lane_reduce}{Choose one from : \code{'2_to_1'}, \code{'3_to_2'}, \code{'4_to_3'}}
}
\description{
It follows <Table 8-20> in KHCM(2013)
}
\examples{
CRF('court_board')
CRF('lane_reduce', '3_to_2')
}
