% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{d}
\alias{d}
\title{Average control delay per vehicle (d, sec/set)}
\usage{
d(d_1 = NULL, d_2 = NULL, d_3 = NULL, PF = NULL, f_cw = 1)
}
\arguments{
\item{f_cw}{Pedestrian crossing signal correction factor between signal intersections. See f_cw().}
}
\description{
It follows <Formula 8-43>, <Formula 12-2> in KHCM(2013) p.538
}
