% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_1.R
\name{d_1}
\alias{d_1}
\title{Uniform delay(d_1)}
\usage{
d_1(
  type = NULL,
  Q_b = NULL,
  C = NULL,
  g = NULL,
  X = NULL,
  R = NULL,
  y = NULL,
  T = NULL,
  S = NULL
)
}
\arguments{
\item{type}{*Categorical* Choose one from : \code{1}, \code{2}, \code{3}}

\item{Q_b}{Initial number of waiting vehicles (units)}

\item{C}{Signal period(seconds)}

\item{g}{Effective green time allocated to the relevant lane group (seconds)}

\item{X}{The saturation of the lane group}

\item{R}{Red signal time (sec)}

\item{y}{Flow rate(= V/S)}

\item{T}{Analysis period length (hours)}

\item{S}{Saturated traffic volume of the lane group (vphg)}
}
\description{
It follows <Formula 8-47>, <Formula 8-48>, <Formula 8-49>, <Formula 12-3> in KHCM(2013)
}
