% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_2.R
\name{d_2}
\alias{d_2}
\title{Incremental delay(d_2)}
\usage{
d_2(c = NULL, t = NULL, X = NULL)
}
\arguments{
\item{c}{The capacity (vph) of the corresponding vehicle group during the analysis period}

\item{t}{Analysis period length (hours)}

\item{X}{The saturation of the lane group}
}
\description{
It follows <Formula 8-50>, <Formula 12-4> in KHCM(2013)
}
