% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_3.R
\name{d_3}
\alias{d_3}
\title{Additional delays (delays due to oversaturation queues remaining before the analysis period, d_3)}
\usage{
d_3(type = NULL, Q_b = NULL, c = NULL, V = NULL, T = NULL)
}
\arguments{
\item{type}{Choose one from : \code{1}, \code{2}, \code{3}}

\item{Q_b}{Initial number of vehicles on standby at the beginning of the analysis period (T) (units)}

\item{c}{The capacity (vph) of the corresponding vehicle group during the analysis period}

\item{V}{Arrival traffic volume (vph) for the corresponding lane group during the analysis period}

\item{T}{Analysis period length (hours)}

\item{X}{The saturation of the lane group}
}
\description{
It follows <Formula 8-51>, <Formula 8-52>, <Formula 8-53> in KHCM(2013)
}
