% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_avg_weaving.R
\name{D_avg_weaving}
\alias{D_avg_weaving}
\title{Average Density in Freeway Weaving Section(D_avg_weaving, pcpkmpl)}
\usage{
D_avg_weaving(
  V = NULL,
  V_w = NULL,
  V_nw = NULL,
  S_w = NULL,
  S_nw = NULL,
  N = NULL
)
}
\arguments{
\item{V}{\emph{Numeric} Total traffic volume of the weaving section(pcph). Input values should be transformed to peak-hour's(including heavy vehicle's influence)}

\item{V_w}{\emph{Numeric} Weaving traffic volume(pcph) Input values should be transformed to peak-hour's(including heavy vehicle's influence)}

\item{V_nw}{\emph{Numeric} Non-weaving traffic volume(pcph) Input values should be transformed to peak-hour's(including heavy vehicle's influence)}

\item{S_w}{\emph{Numeric} Average Speed of Weaving Traffic Flow in Freeway Weaving Section(S_w, kph)}

\item{S_nw}{\emph{Numeric} Average Speed of Non-Weaving Traffic Flow in Freeway Weaving Section(S_nw, kph)}

\item{N}{\emph{Numeric} Total number of lanes in the weaving section}
}
\description{
This function calculates the average density in freeway weaving section.(pcpkmpl)
}
\examples{
D_avg_weaving(V = 1000, V_w = 300, V_nw = 700, S_w = 80, S_nw = 75, N = 4)
D_avg_weaving(1100, 350, 750, 85, 80, 6)
}
