% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_g.R
\name{d_g}
\alias{d_g}
\title{Green time erasable distance (d_g, m)}
\usage{
d_g(d_fb = NULL, g_i = NULL, h_bar = NULL)
}
\arguments{
\item{d_fb}{Headspace spacing when the vehicle is stopped (m)}

\item{g_i}{Effective green time (s)}

\item{h_bar}{Saturation difference time (s)}
}
\description{
It follows <Formula 8-59> in KHCM(2013)
}
