% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_A.R, R/d_I.R
\name{d_I}
\alias{d_I}
\title{Average control delay per vehicle on approach A (d_A, sec/unit)}
\usage{
d_I(d_i = NULL, V_i = NULL)

d_I(d_i = NULL, V_i = NULL)
}
\arguments{
\item{d_i}{Average control delay per vehicle of the type point movement flow i(sec/unit)}

\item{V_i}{Corrected traffic volume (vph) of the moving flow i of the end point}

\item{d_A}{Average control delay per vehicle in approach A lane group (seconds/vehicle)}

\item{V_A}{Corrected traffic volume (vph) for the i lane group}
}
\description{
It follows <Formula 8-56> in KHCM(2013)

It follows <Formula 9-10> in KHCM(2013) p.434
}
\examples{
d_I(d_A = c(10, 2.4, 3.5, 6.2), V_A = c(230, 432, 581, 398))
}
