% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_nonsignaled_intersection.R
\name{d_nonsignaled_intersection}
\alias{d_nonsignaled_intersection}
\title{Operation delay at non-signaled intersections(d_nonsignaled_intersection, seconds/unit)}
\usage{
d_nonsignaled_intersection(V_x = NULL, c_m_x = NULL, T = NULL)
}
\arguments{
\item{V_x}{Traffic flow rate for moving flow x (vph)}

\item{c_m_x}{Capacity of flow x (vph)}

\item{T}{Analysis time period (hours) (T = 0.25 means 15 minutes analysis time)}
}
\description{
This function follows <Formula 10-5> in KHCM(2013) p.470
}
