% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDHV.R
\name{DDHV}
\alias{DDHV}
\title{Directional Design Hourly Volume(DDHV)}
\usage{
DDHV(AADT = NULL, region = NULL, road = NULL, lane = NULL)
}
\arguments{
\item{AADT}{\emph{Numeric} Average Annual Daily Traffic(pc/day)}

\item{region}{\emph{Categorical} Choose one from : \code{'city'}, \code{'rural'}}

\item{road}{\emph{Categorical} Choose one from : \code{'general'}, \code{'highway'}}

\item{lane}{\emph{Numeric} The number of lane(round-trip). It should \code{2} or \code{4} or more.}
}
\description{
This function calculate Directional Design Hourly Volume(DDHV).
}
\examples{
DDHV(AADT = 1500, region = 'city', road = 'general', lane = 2)
DDHV(3000, 'city', 'general', 8)
DDHV(1700, 'city', 'general', 6)
}
\keyword{AADT}
\keyword{DDHV}
\keyword{DHV}
\keyword{Design}
\keyword{Directional}
\keyword{Factor}
\keyword{Hourly}
