% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHV.R
\name{DHV}
\alias{DHV}
\title{Design Hour Volume(DHV)}
\usage{
DHV(AADT = NULL, region = NULL, road = NULL, lane = NULL)
}
\arguments{
\item{AADT}{*Numeric* Average Annual Daily Traffic(pc/day, vph)}

\item{region}{*Categorical* Choose one from : \code{'city'}, \code{'rural'}, \code{'tourist_area'}}

\item{road}{*Categorical* Choose one from : \code{'general'}, \code{'highway'}}

\item{lane}{*Numeric* The number of lane(round-trip). It should \code{2} or \code{4} or more.}
}
\description{
This function calculate Design Hour Volume(DHV).
}
\examples{
DHV(AADT = 2000, region = 'tourist_area', road = 'highway', lane = 8)
DHV(AADT = 1500, region = 'city', road = 'general', lane = 2)
DHV(3000, 'city', 'general', 8)
}
\keyword{AADT}
\keyword{DHV}
\keyword{Design}
\keyword{Factor}
\keyword{Hour}
