% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_1.R
\name{E_1}
\alias{E_1}
\title{The straight-line conversion factor of the left turn itself(E_1)}
\usage{
E_1(
  left_signal_operation = NULL,
  left_turn = NULL,
  N_left = NULL,
  N = NULL,
  P = NULL,
  V_o = NULL,
  V_L = NULL,
  V_TH = NULL,
  C = NULL,
  g_c_ratio = NULL
)
}
\arguments{
\item{left_signal_operation}{*Categorical* \code{'two_sided_protection'}, \code{'direct_left_sinultaneous'}, \code{'unprotected'}}

\item{left_turn}{*Categorical* \code{'private'}, \code{'public'}}

\item{N_left}{*categorical* Number of left-turn lane. \code{1}, \code{2}}

\item{N}{*Numeric* Number of access lanes (excluding dedicated left-turn lanes)}

\item{P}{*Numeric* Number of possible unprotected left turns per gap by opposite traffic volume}

\item{V_o}{*Numeric* Opposite traffic volume(vph)}

\item{V_L}{*Numeric* Left turn traffic volume(vph)}

\item{V_TH}{*Numeric* Straight-through traffic volume(vph)}

\item{C}{*Numeric* Period(s)}

\item{g_c_ratio}{*Numeric* Effective green hour rate}
}
\description{
This function follows <Formula 8-6> in KHCM(2013)
}
\examples{
E_1(left_signal_operation = 'two_sided_protection', left_turn = 'private', N_left = 2)
E_1(left_signal_operation = 'direct_left_simultaneous', left_turn = 'public', N_left = 1)
E_1(left_signal_operation = 'unprotected', left_turn = 'public', N = 2, P = 10, V_o = 100, V_L = 50, V_TH = 280, C = 120, g_c_ratio = 0.2)
}
