% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_L.R
\name{E_L}
\alias{E_L}
\title{Comprehensive straight-line conversion factor for left-turning lanes(E_L)}
\usage{
E_L(E_1 = NULL, E_p = NULL, E_u = NULL)
}
\arguments{
\item{E_1}{Influence of left turn itself according to the number of left turn lanes}

\item{E_p}{Influence of curve radius on left turn trajectory}

\item{E_u}{Influence of u-tern}
}
\description{
This function follows <Formula 8-6> in KHCM(2013)
}
