% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_p.R
\name{E_p}
\alias{E_p}
\title{Straight conversion factor by left turn curve radius(E_p)}
\usage{
E_p(radius = NULL)
}
\arguments{
\item{radius}{*Numeric* Left turn curve radius(m)}
}
\description{
This function follows <Table 8-9> in KHCM(2013)
}
\examples{
E_p(8.2)
}
