% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_R.R
\name{E_R}
\alias{E_R}
\title{Percentage of crosswalk signal time for which a right turn is not available (E_R)}
\usage{
E_R(
  dowry = NULL,
  V_R = NULL,
  V_TH = NULL,
  f_c = NULL,
  G_p = NULL,
  C = NULL,
  L_H = NULL,
  N = NULL
)
}
\arguments{
\item{dowry}{\emph{Categorical} Choose one from : \code{'yes'}, \code{'no'}}

\item{V_R}{\emph{Numeric} Corrected Right Turn Traffic (vph)}

\item{V_TH}{\emph{Numeric} straight-through traffic (vph)}

\item{f_c}{\emph{Numeric} Proportion of time to obstruct right turns among pedestrian crossing signals}

\item{G_p}{\emph{Numeric} Pedestrian crossing signal (sec.)}

\item{C}{\emph{Numeric} Signal Cycle(seconds)}

\item{L_H}{\emph{Numeric} Roadside friction (seconds) due to entry and exit of the back road, bus stop, and on-street parking}

\item{N}{\emph{Categorical} Number of lanes. Choose one from : \code{1}, \code{2}, \code{3}, \code{4}, \code{5}, \code{6}}
}
\description{
This function follows <Table 8-13> in KHCM(2013)
}
\examples{
E_R(dowry = 'yes', V_R = 320, V_TH = 280, f_c = 0.2, G_p = 40, C = 200, L_H = 30.22, N = 4)
}
