% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_T_2lane_road.R
\name{E_T_2lane_Road}
\alias{E_T_2lane_Road}
\title{Heavy Vehicle Correction Factor according to the Traffic Volume in the Direction(E_T)}
\usage{
E_T_2lane_Road(
  topography = NULL,
  gradient = NULL,
  L = NULL,
  V_d = NULL,
  output = NULL
)
}
\arguments{
\item{topography}{*Categorical*  Choose one from : \code{'flatland'}, \code{'hill'}, \code{'specific_slope}}

\item{gradient}{*Numeric* If \code{topography = 'specific_slope'}, must use it. The gradient of slope.}

\item{L}{*Numeric* If \code{topography = 'specific_slope'}, must use it. The length of the slope.(m)}

\item{output}{*Categorical* Choose one from : \code{'travel_speed'}, \code{'TDR'}}

\item{V}{*Numeric* If \code{topography = 'specific_slope'}, must use it. Traffic Volume in direction(pcph?)}
}
\description{
This function follows <Table 7-4>, <Table 7-5>, <Table 7-6>, <Table 7-7> (p.177)
}
\examples{
E_T_2lane_Road(topography = 'flatland', V_d = 332, output = 'TDR')
E_T_2lane_Road(topography = 'hill', V_d = 459, output = 'travel_speed')
E_T_2lane_Road(topography = 'specific_slope', gradient = 0.04, L = 1200, V_d = 838, output = 'travel_speed')
}
