% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_T_roundabout.R
\name{E_T_roundabout}
\alias{E_T_roundabout}
\title{Conversion factor for passenger cars of heavy vehicles(E_T_roundabout)}
\usage{
E_T_roundabout(lane = NULL, hv_ratio = NULL)
}
\arguments{
\item{lane}{Choose one from L \code{1}, \code{2}}

\item{hv_ratio}{Heavy Vehicle Ratio(\%)}
}
\description{
It follows <Table 11-4> in KHCM(2013) p.500
}
