% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_u.R
\name{E_u}
\alias{E_u}
\title{Forward conversion factor of left turn by U-turn %(E_u)}
\usage{
E_u(N_left = NULL, u_tern_ratio = NULL)
}
\arguments{
\item{N_left}{*Categorical* Number of left-turn lane. \code{1}, \code{2}}

\item{u_tern_ratio}{*Numeric* U-tern ratio(%)}
}
\description{
This function follows <Table 8-10>, <Table 8-11> in KHCM(2013)
}
\examples{
E_u(1, 32)
E_u(2, 42.5)
}
