% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESL_2lane_road.R
\name{ESL_2lane_road}
\alias{ESL_2lane_road}
\title{Signal Intersection Length of Sphere of Influence(ESL, m)}
\usage{
ESL_2lane_road(V_d = NULL, LB = NULL, g_c_ratio = NULL, L = NULL, DIS = NULL)
}
\arguments{
\item{V_d}{\emph{Numeric} Traffic in the direction of travel(pcph?)}

\item{LB}{\emph{Categorical} Whether the left turn-only lane exist. \code{LB = 1} means exist. \code{LB = 0} means not.}

\item{g_c_ratio}{\emph{Numeric} Effective green time ratio}

\item{L}{\emph{Numeric} Direction Traffic Volume Left Turn Ratio}

\item{DIS}{\emph{Numeric} Direction traffic volume straight forward ratio}
}
\description{
This function follows <Formula 7-3>
}
\examples{
ESL_2lane_road(V_d = 1000, LB = 1, g_c_ratio = 0.23, L = 0.2, DIS = 0.73)
ESL_2lane_road(300, 0, 0.11, 0.123, 0.84)
}
