% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_A.R
\name{F_A}
\alias{F_A}
\title{Speed Correction Factor from Number of inlet and outlet points(F_A, kph)}
\usage{
F_A(n_inlet_outlet = 0)
}
\arguments{
\item{n_inlet_outlet}{Number of inlet and outlet points.(ea/km)}
}
\description{
It follows <Table 6-5> in KHCM(2013).
}
\examples{
F(n_inlet_outlet = 10)
F_A(8)
}
\keyword{inlet}
\keyword{outlet}
\keyword{point}
\keyword{reduced}
\keyword{speed}
