% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_B.R
\name{F_B}
\alias{F_B}
\title{Bendiness Speed Correction Factor(F_B)}
\usage{
F_B(B = NULL, speed = NULL)
}
\arguments{
\item{B}{\emph{Numeric} Bendiness. Horizontal Curve Curvature in Multi-lane Road(B, degree/km)}

\item{speed}{\emph{Numeric} Speed(kph). What speed?}
}
\description{
This function calculates horizontal curve curvature in multi-lane road. It follows <Table 6-3> in KHCM(2013).
}
\examples{
F_B(B = 4, speed = 70)
F_B(30, 87)
}
\keyword{Density}
\keyword{LOS}
\keyword{Level}
\keyword{Service}
\keyword{V/C}
\keyword{of}
\keyword{ratio}
