% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_c.R
\name{f_c}
\alias{f_c}
\title{Percentage of crosswalk signal time for which a right turn is not available (f_c)}
\usage{
f_c(N_ped = NULL)
}
\arguments{
\item{N_ped}{Number of pedestrians crossing (in both directions, person/hour)}
}
\description{
This function follows <Table 8-13> in KHCM(2013)
}
\examples{
f_c(2883)
}
