% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_cw.R
\name{f_cw}
\alias{f_cw}
\title{Pedestrian crossing signal correction factor between signal intersections(f_cw)}
\usage{
f_cw(interlock = NULL, n_crossing = NULL)
}
\arguments{
\item{interlock}{Choose one from: \code{'yes'}, \code{'no'}}

\item{n_crossing}{Number of crosswalk road in target section}
}
\description{
This function follows <Table 12-7> in KHCM(2013) p.539
}
\examples{
f_cw('yes', 4)
}
