% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_dk.R
\name{f_dk}
\alias{f_dk}
\title{Capacity Calculation Correction Factor for The Night Situation of The Basic Section of The Freeway(f_dk)}
\usage{
f_dk(design_speed = NULL, day_night = NULL)
}
\arguments{
\item{design_speed}{\emph{Categorical} Choos one from : \code{120}, \code{100}, \code{80}}

\item{day_night}{\emph{Categorical} Choose one from : \code{'day'}, \code{'night'}}
}
\description{
This function decides the capacity calculation factor for the basic section of the freeway(f_dk).
}
\examples{
f_dk(design_speed = 100, 'day')
f_dk(design_speed = 120, 'night')
f_dk(80, 'night')
}
