% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_g.R
\name{f_g}
\alias{f_g}
\title{Inclination correction factor(f_g)}
\usage{
f_g(gradient = NULL)
}
\arguments{
\item{gradient}{%}
}
\description{
It follows <Table 8-16> in KHCM(2013)
}
\examples{
f_g(gradient = 2.72)
f_g(3.945)
}
