% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_H.R
\name{F_H}
\alias{F_H}
\title{Speed Correction Factor from Hillness(F_H)}
\usage{
F_H(H = NULL, vehicle_type = NULL, speed = NULL)
}
\arguments{
\item{H}{*Numeric* Hillness. Vertical Curve Curvature in Multi-lane Road(H, degree/km)}

\item{vehicle_type}{*Categorical* Choose one from : \code{'normal_vehicle'}, \code{'heavy_vehicle'}}

\item{speed}{*Numeric* Speed(kph). What speed?}
}
\description{
This function calculates vertical curve curvature in multi-lane road. It follows <Table 6-4> in KHCM(2013).
}
\examples{
F_H(H = 30.2, vehicle_type = 'heavy_vehicle', speed = 70)
F_H(23, 'normal_vehicle', 87)
}
\keyword{Density}
\keyword{LOS}
\keyword{Level}
\keyword{Service}
\keyword{V/C}
\keyword{of}
\keyword{ratio}
