% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_hv.R
\name{f_hv}
\alias{f_hv}
\title{Heavy Vehicle Factors(f_hv)}
\usage{
f_hv(
  topography = NULL,
  prop_small_hv = 0,
  prop_middle_hv = 0,
  prop_large_hv = 0,
  prop_hv = 0,
  slope = 0,
  slope_length = 0
)
}
\arguments{
\item{topography}{*String* Choose one from : \code{'flatland'}, \code{'hill'}, \code{'mountain'}, \code{'specific_slope}}

\item{prop_small_hv}{*Numeric* The ratio of small heavy vehicle composition(Trucks less than 2.5 tons, buses less than 16 passengers)}

\item{prop_middle_hv}{*Numeric* The ratio of middle heavy vehicle composition(Trucks of 2.5 tons or more, buses with 16 passengers or more)}

\item{prop_large_hv}{*Numeric* The ratio of large heavy vehicle composition(Semi trailer or full trailer)}

\item{prop_hv}{*Numeric* The ratio of heavy vehicles.}

\item{slope}{*Numeric* Slope rate(%)}

\item{slope_length}{*Numeric* The length of the slope(km)}
}
\description{
This function calculates Heavy Vehicle Factors(f_hv) by geographical features in basic freeway section.
}
\examples{
f_hv(topography = 'flatland', prop_small_hv = 0.3, prop_middle_hv = 0.1, prop_large_hv = 0.1
f_hv(topography = 'hill', prop_small_hv = 0.11, prop_middle_hv = 0.05, prop_large_hv = 0.03)
f_hv('mountain', 0.18, 0.14, 0)
f_hv('specific_slope', slope = 3.5, slope_length = 2, prop_hv = 0.3)
}
