% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_hv3.R
\name{f_hv3}
\alias{f_hv3}
\title{Heavy Vehicle Factors in Signalized Intersection(f_hv3)}
\usage{
f_hv3(P = NULL)
}
\arguments{
\item{P}{Mixing ratio of heavy vehicles to actual traffic volume}
}
\description{
It follows <Formula 8-39> in KHCM(2013)
}
