% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_hv_roundabout.R
\name{f_hv_roundabout}
\alias{f_hv_roundabout}
\title{Heavy Vehicle Factors in Roundabout(f_hv_roundabout)}
\usage{
f_hv_roundabout(lane = NULL, hv_ratio = NULL, P_T = NULL)
}
\arguments{
\item{lane}{Choose one from L \code{1}, \code{2}}

\item{hv_ratio}{Heavy Vehicle Ratio(%)}

\item{P_T}{Heavy Vehicle Mixing Ratio}
}
\description{
It follows <Formula 11-4> in KHCM(2013) p.500
}
