% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_iw.R
\name{f_iw}
\alias{f_iw}
\title{Capacity Calculation Correction Factor for Bad Weather Conditions in The Basic Section of The Freeway(f_iw)}
\usage{
f_iw(design_speed = NULL, weather = NULL, precipitation = NULL)
}
\arguments{
\item{design_speed}{\emph{Categorical} Choos one from : \code{120}, \code{100}, \code{80}}

\item{weather}{\emph{Categorical} Choose one from : \code{'sunny'}, \code{'rainy'}, \code{'snowy'}}

\item{precipitation}{\emph{Numeric} Rainfall or snowfall or hail (mm/h)}
}
\description{
This function decides the capacity calculation factor for the basic section of the freeway(f_dk).
}
\examples{
f_iw(design_speed = 80, weather = 'sunny')
f_iw(design_speed = 100, weather = 'rainy', precipitation = 8)
f_iw(120, 'snowy', 12)
}
