% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_iw3.R
\name{f_iw3}
\alias{f_iw3}
\title{Capacity calculation correction factor for bad weather conditions at signal intersections(f_iw3)}
\usage{
f_iw3(weather = NULL, precipitation = NULL)
}
\arguments{
\item{weather}{\emph{Categorical} Choose one from : \code{'sunny'}, \code{'rainy'}, \code{'snowy'}}

\item{precipitation}{\emph{Numeric} Rainfall or snowfall or hail (mm/h)}
}
\description{
It follows <Table 8-28> in KHCM(2013)
}
\examples{
f_iw3('sunny')
f_iw3('rainy', 32.1)
f_iw3('snowy', 8.2334)
}
