% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_LT.R
\name{f_LT}
\alias{f_LT}
\title{Left turn correction factor(f_LT)}
\usage{
f_LT(lane_type = NULL, E_L = NULL, P_L = NULL, P_LT = NULL)
}
\arguments{
\item{lane_type}{\emph{Categorical} Choose from one : \code{'dedicated'}, \code{'practical_only'}, \code{'public'}}
}
\description{
This function follows <Formula 8-27>, <Formula 8-28>, <Formula 8-30> in KHCM(2013)
}
