% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_LT_RT.R
\name{f_LT_RT}
\alias{f_LT_RT}
\title{Turn correction factor in integrated lane(f_LT_RT)}
\usage{
f_LT_RT(E_L = NULL, E_R = NULL, P_LT = NULL, P_RT = NULL)
}
\description{
This function follows <Formula 8-32> in KHCM(2013)
}
