% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_np.R
\name{f_np}
\alias{f_np}
\title{Correction Factor According to Distribution by Direction and Ratio of Overtaking Prohibited Section(f_np_ATS or f_np_D)}
\usage{
f_np(V_d = NULL, V_o = NULL, overtaking_prop = NULL, output = NULL)
}
\arguments{
\item{V_d}{\emph{Numeric} One-way traffic volume(pcph?)}

\item{V_o}{\emph{Numeric} Opposite traffic volume(pcph?)}

\item{overtaking_prop}{\emph{Numeric} Proportion of no overtaking section.}

\item{output}{\emph{Categorical} Choose one from : \code{'travel_speed'}(kph), \code{'TDR'}}
}
\description{
It follows <Table 7-12> in KHCM(2013).
}
\examples{
f_np(V_d = 800, V_o = 600, overtaking_prop = 0.4, output = 'travel_speed')
f_np(1000, 1200, 0.2, 'TDR')
}
