% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_p.R
\name{f_p}
\alias{f_p}
\title{Influence coefficient on the number of pedestrians by conflicting traffic volume(f_p)}
\usage{
f_p(lane = NULL, V_ci = NULL, p_v = NULL)
}
\arguments{
\item{lane}{roundabout type. Choose one from : \code{1}, \code{2}}

\item{V_ci}{Conflict Traffic Volume(pcph)}

\item{p_v}{Amount of passenger (person/hour)}
}
\description{
It follows <Table 11-3> in KHCM(2013) p.498
}
\examples{
f_p(1, 1232, 323)
f_p(2, 546, 215)
}
