% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_pl.R
\name{f_pl}
\alias{f_pl}
\title{Correction Coefficient of Total Delay in 2+1 Lane Road Section(f_pl)}
\usage{
f_pl(V = NULL)
}
\arguments{
\item{V}{*Numeric* Traffic Volume(pcphpl)}
}
\description{
This function follows <Table 7-14>
}
\examples{
f_pl(V = 1391)
f_pl(999)
}
