% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_R.R
\name{F_R}
\alias{F_R}
\title{Right Turn Traffic Volume Correction Factor(F_R)}
\usage{
F_R(intersection_prongs = NULL, dowry = NULL)
}
\arguments{
\item{intersection_prongs}{\emph{Categorical} Choose one from : \code{3}, \code{4}}

\item{dowry}{\emph{Categorical} \code{'yes'}, \code{'no'}}
}
\description{
This function follows <Table 8-6> in KHCM(2013)
}
\examples{
F_R(3)
F_R(4, 'no')
}
\keyword{Coefficient}
\keyword{Correction}
\keyword{Factor}
\keyword{Right}
\keyword{Traffic}
\keyword{Turn}
\keyword{Volume}
