% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RT.R
\name{f_RT}
\alias{f_RT}
\title{Right turn correction factor(f_RT)}
\usage{
f_RT(lane_type = NULL, E_R = NULL, P_R = NULL, P_RT = NULL)
}
\arguments{
\item{lane_type}{*Categorical* Choose from one : \code{'practical_only'}, \code{'public'}}
}
\description{
This function follows <Formula 8-29>, <Formula 8-31> in KHCM(2013)
}
