% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_S.R
\name{F_S}
\alias{F_S}
\title{Speed Correction Factor from Traffic Light(F_S, kph)}
\usage{
F_S(
  V = NULL,
  g_c_ratio = NULL,
  traffic_light_density = NULL,
  vehicle_type = NULL
)
}
\arguments{
\item{V}{*Numeric* Traffic Volume(vphpl). This factor is available when traffic volume is under than 500vphpl.}

\item{g_c_ratio}{*Numeric* Average of the signal going straight in the main direction.}

\item{traffic_light_density}{*Numeric* Density of traffic lights(ea/km)}

\item{vehicle_type}{*Categorical* Choose one from : \code{'normal_vehicle'}, \code{'heavy_vehicle'}}
}
\description{
It follows <Table 6-6> in KHCM(2013).
}
\examples{
F_S(V = 499, g_c_ratio = 0.80, traffic_light_density = 0.2, vehicle_type = 'normal_vehicle')
F_S(300, 0.7, 0.4, 'heavy_vehicle')
}
\keyword{correction}
\keyword{factor}
\keyword{light}
\keyword{speed}
\keyword{traffic}
