% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_U.R
\name{F_U}
\alias{F_U}
\title{Lane Use Coefficient(F_U)}
\usage{
F_U(V_avg = NULL, design_level = NULL, N = NULL)
}
\arguments{
\item{V_avg}{\emph{Numeric} Average traffic volume by lane.(vphpl)}

\item{design_level}{\emph{Categorical} Choose one from : \code{C}, \code{D}, \code{E}}

\item{N}{\emph{Numeric} Number of exclusive lanes going straight}
}
\description{
This function decides Level of Service(LOS). It follows <Table 8-5>
}
\examples{
F_U(V_avg = 1200, N = 3)
F_U(790, 2)
}
\keyword{Coefficient}
\keyword{Lane}
\keyword{Use}
