% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ub.R
\name{f_ub}
\alias{f_ub}
\title{Upstream stop influence correction factor(f_ub)}
\usage{
f_ub(sep_l = NULL)
}
\arguments{
\item{sep_l}{The separation distance of the upstream stop(m)}
}
\description{
It follows <Table 8-19> in KHCM(2013)
}
\examples{
f_ub(112.45)
}
