% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_V.R
\name{F_V}
\alias{F_V}
\title{Speed Correction Factor from Traffic Volume(F_V, kph)}
\usage{
F_V(V = NULL, vehicle_type = NULL, traffic_light_density = NULL)
}
\arguments{
\item{V}{*Numeric* Traffic Volume(vphpl). This factor is available when traffic volume is under than 500vphpl.}

\item{vehicle_type}{*Categorical* Choose one from : \code{'normal_vehicle'}, \code{'heavy_vehicle'}}

\item{traffic_light_density}{*Numeric* Density of traffic lights(ea/km)}
}
\description{
It follows <Table 6-7> in KHCM(2013).
}
\examples{
F_V(V = 1231, vehicle_type = 'normal_vehicle', traffic_light_density = 0.45)
F_V(999, 'heavy_vehicle', 0.3293)
}
\keyword{correction}
\keyword{factor}
\keyword{light}
\keyword{speed}
\keyword{traffic}
