% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_w.R
\name{f_w}
\alias{f_w}
\title{The One Side Lane Width and Lateral Clearance Factor(f_w)}
\usage{
f_w(obstacle = NULL, lateral_clearance = NULL, lane = NULL, lane_width = NULL)
}
\arguments{
\item{obstacle}{\emph{Categorical} Choose one from :\code{'one_side'}, \code{'both_sides'}}

\item{lateral_clearance}{\emph{Numeric} The lateral clearance of lane. From the center line of the center line to the pavement edge of the lane. The Average of left lateral clearance and right lateral clearance.}

\item{lane}{\emph{Numeric} The number of freeway lane(Round-trip lane). It must be 4 or more.}

\item{lane_width}{\emph{Numeric} The width of each lane(m). It must be more than 2.75}
}
\description{
This function calculates the one side lane width and lateral clearance factor(f_w) in given conditions.
}
\examples{
f_w(obstacle = 'one_side', lateral_clearance = 1.4, lane = 8, lane_width = 3.1)
f_w(obstacle = 'both_sides', lateral_clearance = 1, lane = 8, lane_width = 3)
}
