% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_w3.R
\name{f_w3}
\alias{f_w3}
\title{Lane width correction factor(f_w3)}
\usage{
f_w3(lane_width = NULL)
}
\arguments{
\item{lane_width}{When the lane widths within a lane group are different, the average value of these is used.}
}
\description{
It follows <Table 8-15> in KHCM(2013)
}
\examples{
f_w3(2.89)
f_w3(lane_width = 3.23)
}
