% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_w_ATS.R
\name{f_w_ATS}
\alias{f_w_ATS}
\title{The One Side Lane Width and Lateral Clearance Factor in 2-lane Road(f_w_ATS, kph)}
\usage{
f_w_ATS(lateral_clearance = NULL, lane_width = NULL)
}
\arguments{
\item{lateral_clearance}{*Numeric* The lateral clearance of lane. From the center line of the center line to the pavement edge of the lane. The Average of left lateral clearance and right lateral clearance.}

\item{lane_width}{*Numeric* The width of each lane(m). It must be more than 2.75}
}
\description{
This function maximum reduced speed. It follows <Table 7-11> in KHCM(2013).
}
\examples{
f_w_ATS(lateral_clearance = 1.0, lane_width = 3.134)
f_w_ATS(0.5, 3.6478)
}
