% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_w_D.R
\name{f_w_D}
\alias{f_w_D}
\title{The One Side Lane Width and Lateral Clearance Factor in 2-lane Road(f_w_D, \%)}
\usage{
f_w_D(lateral_clearance = NULL, lane_width = NULL)
}
\arguments{
\item{lateral_clearance}{\emph{Numeric} The lateral clearance of lane. From the center line of the center line to the pavement edge of the lane. The Average of left lateral clearance and right lateral clearance.}

\item{lane_width}{\emph{Numeric} The width of each lane(m). It must be more than 2.75}
}
\description{
This function calculates the one side lane width and lateral clearance factor in 2-lane road(f_w_D) in given conditions.
}
\examples{
f_w_D(lateral_clearance = 2.3, lane_width = 3.19)
f_w_D(2.99, 2.83)
}
