% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_wc.R
\name{F_wc}
\alias{F_wc}
\title{The One Side Lane Width and Lateral Clearance Speed Correction Factor : Reduced Maximum Travel Speed in Multi-lane Road(F_wc, kph)}
\usage{
F_wc(lateral_clearance = NULL, lane_width = NULL)
}
\arguments{
\item{lateral_clearance}{\emph{Numeric} The lateral clearance of lane. From the center line of the center line to the pavement edge of the lane. The Average of left lateral clearance and right lateral clearance.}

\item{lane_width}{\emph{Categorical} The width of each lane(m). Choose one from: \code{3.5}, \code{3.25}, \code{3}}
}
\description{
This function calculates the one side lane width and lateral clearance speed correction factor(F_wc) in given conditions. It follows <Table 6-2> in KHCM(2013).
}
\examples{
F_wc(lateral_clearance = 2.2, lane_width = 3.5)
F_wc(1.9, 3.25)
}
