% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_WZ.R
\name{f_WZ}
\alias{f_WZ}
\title{Saturated traffic flow rate correction formula and coefficients for road occupancy construction sections in the upper and downstream parts of the signal intersection(f_WZ)}
\usage{
f_WZ(
  d_g = NULL,
  N = NULL,
  N_WZ = NULL,
  d_WZ = NULL,
  L_WZ = NULL,
  g_T = NULL,
  CRF = NULL,
  g_i = NULL
)
}
\arguments{
\item{d_g}{During green time, the distance that the queue upstream of the stop line can be erased (m)}

\item{N}{Number of lanes (number of stop-line lanes excluding left-turn pockets)}

\item{N_WZ}{The number of closed lanes under the road occupation construction (measured in the cross section of the construction section)}

\item{d_WZ}{Construction section separation distance (m)}

\item{L_WZ}{Construction section length (m)}

\item{g_T}{Total shock wave duration (s)}

\item{CRF}{Capacity Reduction Index}

\item{g_i}{Effective green time (s)}

\item{construction_loc}{Upstream or downstream of the stop line. Choose one from : \code{'upstream'}, \code{'downstream'}}
}
\description{
It follows <Table 8-22> in KHCM(2013)
}
