% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_speed_arterial.R
\name{free_speed_arterial}
\alias{free_speed_arterial}
\title{Free speed on arterial roads(free_speed_arterial)}
\usage{
free_speed_arterial(road_standard = NULL, road_condition = NULL)
}
\arguments{
\item{road_standard}{\code{'high'}, \code{'intermediate'}, \code{'low'}}

\item{road_condition}{\code{'good'}, \code{'normal'}}
}
\description{
It follows <Table 12-2> in KHCM(2013) p.533
}
\examples{
free_speed_arterial('high', 'good')
}
