% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_i_backtick_downstream.R
\name{g_i_backtick_downstream}
\alias{g_i_backtick_downstream}
\title{Corrected effective green time at the downstream intersection reflecting the additional green loss time in the downstream part due to non-use of green time(g_i_backtick_downstream)}
\usage{
g_i_backtick_downstream(G = NULL, G_L = NULL, L_B = NULL, L_DS = NULL)
}
\arguments{
\item{G}{Green time (seconds)}

\item{G_L}{Progress extension time (=2.0 seconds)}

\item{L_B}{Departure loss time (=2.3 seconds)}

\item{L_DS}{Additional green loss time due to non-use of green time (seconds)}
}
\description{
It follows <Formula 9-8> in KHCM(2013), p.433
}
