% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_i_backtick_upstream.R
\name{g_i_backtick_upstream}
\alias{g_i_backtick_upstream}
\title{Corrected effective green time at the upstream intersection reflecting the loss time due to the waiting vehicle of the internal link(g_i_backtick)}
\usage{
g_i_backtick_upstream(G = NULL, L_B = NULL, G_L = NULL, L_Q = NULL)
}
\arguments{
\item{G}{Green time (seconds)}

\item{L_B}{Departure loss time (=2.3 seconds)}

\item{G_L}{Progress extension time (=2.0 seconds)}

\item{L_Q}{Additional green loss time (sec) due to the initial queue of the internal link}
}
\description{
It follows <Formula 9-7> in KHCM(2013), p.432
}
