% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_T.R
\name{g_T}
\alias{g_T}
\title{Total shock wave session time(g_T)}
\usage{
g_T(d_T = NULL, d_WZ = NULL, d_IW = NULL, u_o = NULL, w = NULL)
}
\arguments{
\item{d_T}{Construction section taper length (m)}

\item{d_WZ}{Construction section separation distance (m)}

\item{d_IW}{Intersection width (m)}

\item{u_o}{Average moving speed(m/s)}

\item{w}{Shock wave speed(m/s)}
}
\description{
It follows <Formula 8-60>, <Formula 8-61> in KHCM(2013)
}
