% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_multilane_road.R
\name{H_multilane_road}
\alias{H_multilane_road}
\title{Hillness in Multi-lane Road(H, m/km)}
\usage{
H_multilane_road(diff_in_height = NULL, L = NULL)
}
\arguments{
\item{diff_in_height}{\emph{Series} Series of each difference in height on direction. \code{c(height1, height2, height3, ...)} (m)}

\item{L}{\emph{Numeric} Length of the route(km). Generally, \code{L = 3}}
}
\description{
This function calculates difference in height in multi-lane road. It follows <Formula 6-1> in KHCM(2013).
}
\examples{
H_multilane_road(diff_in_height = c(3, 2, 8, 3.2, 5), L = 3)
H_multilane_road(c(2.1, 5.3, 4.48, 3.32, 2.1), 3.4)
}
\keyword{Hillness}
\keyword{Multilane}
\keyword{road}
