% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_capacity_freeway_ramp_junction.R
\name{influence_capacity_freeway_ramp_junction}
\alias{influence_capacity_freeway_ramp_junction}
\title{The Inlet or Outlet Capacity of the Freeway Ramp Junction(pcph)}
\usage{
influence_capacity_freeway_ramp_junction(
  free_speed = NULL,
  V = NULL,
  N = NULL,
  inlet_outlet = NULL
)
}
\arguments{
\item{free_speed}{*Numeric* It means free speed of main lane.(kph)}

\item{V}{*Numeric* Main line traffic volume of branch and confluence section(pcph)}

\item{N}{*Numeric* Total number of lanes(one-way). It must be 2 or more.}

\item{inlet_outlet}{*Categirocal* Choose output. \code{inlet}, \code{outlet}}
}
\description{
This function decides inlet or outlet capacity of the freeway ramp junction(pcph). It follows <Table 4-1> in KHCM(2013).
}
\examples{
influence_capacity_freeway_ramp_junction(free_speed = 120, V = 5000, N = 5, inlet_outlet = 'outlet')
influence_capacity_freeway_ramp_junction(100, 4000, 3, 'inlet')
}
\keyword{capacity}
\keyword{freeway}
\keyword{junction}
\keyword{ramp}
