% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_bb.R
\name{L_bb}
\alias{L_bb}
\title{Interference time due to bus stop(L_bb)}
\usage{
L_bb(T_b = NULL, l = NULL, V_b = NULL)
}
\arguments{
\item{T_b}{Saturation head time increment according to one bus stop (sec). See T_b()}

\item{l}{Distance from stop line to bus stop (m)}

\item{V_b}{Number of bus stops per hour}
}
\description{
This function follows <Formula 8-8> in KHCM(2013)
}
\examples{
L_bb(T_b = 10.8, l = 30, V_b = 12)
L_bb(1.4, 15, 15)
}
