% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_DS.R
\name{L_DS}
\alias{L_DS}
\title{Additional green loss time downstream due to non-use of green time(L_DS)}
\usage{
L_DS(CG_DS = NULL, Q_V = NULL, h_s = NULL)
}
\arguments{
\item{CG_DS}{Period in which the upstream intersection shows a left turn and the downstream intersection shows green}

\item{Q_V}{Average number of waiting vehicles in the internal link at the start of each signal (units)}

\item{h_s}{Saturation difference time (sec/set)}
}
\description{
TIt follows <Formula 9-6> in KHCM(2013), p.429
}
