% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_dw.R
\name{L_dw}
\alias{L_dw}
\title{Interference time by vehicles entering and exiting(L_dw)}
\usage{
L_dw(V_en = NULL, V_ex = NULL)
}
\arguments{
\item{V_en}{Traffic entering the arterial road (vph)}

\item{V_ex}{Traffic leaving the arterial road (vph)}
}
\description{
This function follows <Formula 8-7> in KHCM(2013)
}
\examples{
L_dw(V_en = 800, V_ex = 732)
L_dw(132, 241)
}
