% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_H.R
\name{L_H}
\alias{L_H}
\title{Comprehensive roadside friction(L_H)}
\usage{
L_H(L_dw = NULL, L_bb = NULL, L_p = NULL)
}
\arguments{
\item{L_dw}{Interference time by vehicles entering and exiting.(s) See L_dw()}

\item{L_bb}{Interference time due to bus stop(s). See L_bb()}

\item{L_p}{Interference time due to parking activities(s). See L_p()}
}
\description{
This function follows <Formula 8-10> in KHCM(2013)
}
\examples{
L_H(30.2, 38.22, 101.33)
}
