% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_p.R
\name{L_p}
\alias{L_p}
\title{Interference time due to parking activities(L_p)}
\usage{
L_p(street_parking = NULL, V_park)
}
\arguments{
\item{street_parking}{Choose one from: \code{'allowed'}, \code{'not_allowed'}}

\item{V_park}{Parking activity per hour (vph)}
}
\description{
This function follows <Formula 8-9> in KHCM(2013)
}
\examples{
L_p(street_parking = 'allowed', V_park = 30)
L_p('not_allowed')
}
