% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_Q.R
\name{L_Q}
\alias{L_Q}
\title{Additional green loss time at the upstream intersection due to the queuing of the internal link(L_Q)}
\usage{
L_Q(Q_L = NULL, mu_F = NULL, mu_B = NULL, tau_0 = NULL, tau_1 = NULL)
}
\arguments{
\item{Q_L}{Initial queue length (m) by appearance of internal link between intersections. Q_V * l_Q}

\item{mu_F}{The speed at which the front of the queue moves to the end of the internal link (m/sec)}

\item{mu_B}{The speed at which the end of the queue moves to the end of the internal link (m/sec)}

\item{tau_0}{Time (in seconds) from the start until the first vehicle emitted from the upstream intersection arrives at the end of the queue}

\item{tau_1}{Time (in seconds) from t_0 until the queue arrives at the end of the internal link}
}
\description{
TIt follows <Formula 9-1> in KHCM(2013), p.429
}
