% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_2lane_road.R
\name{LOS_2lane_road}
\alias{LOS_2lane_road}
\title{Level of Service(LOS) in 2-lane Road}
\usage{
LOS_2lane_road(
  type = NULL,
  free_speed = NULL,
  speed = NULL,
  TDR = NULL,
  V = NULL
)
}
\arguments{
\item{type}{*Categorical* Type of the 2-lane road. Choose one from : \code{'type_1'}, \code{'type_2'}}

\item{free_speed}{*Categorical* kph. Choose one from : \code{100}, \code{90}, \code{80}}

\item{speed}{*Numeric* It means speed(kph)}

\item{TDR}{*Numeric* Total delay rate(%). see TDR()}

\item{V}{*Numeric* Traffic Volume(pcph)}
}
\description{
This function decides Level of Service(LOS). It follows <Table 7-2>
}
\examples{
LOS_2lane_road(type = 'type_1', free_speed = 100, speed = 74, TDR = 44, V = 2900)
LOS_2lane_road('type_2', 60, 45, 23, 1200)
}
\keyword{Density}
\keyword{LOS}
\keyword{Level}
\keyword{Service}
\keyword{V/C}
\keyword{of}
\keyword{ratio}
