% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_arterial.R
\name{LOS_arterial}
\alias{LOS_arterial}
\title{Level of Service(LOS) by average travel speed on arterial roads(LOS_arterial)}
\usage{
LOS_arterial(free_speed = NULL, avg_speed = NULL)
}
\arguments{
\item{free_speed}{Choose one from: \code{80}, \code{70}, \code{60}}

\item{avg_speed}{*Numeric* Average travel speed(kph)}
}
\description{
It follows <Table 12-1> in KHCM(2013) p.529
}
\examples{
LOS_arterial(free_speed = 70, avg_speed = 38)
LOS_arterial(80, 49)
}
