% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_bus_interval.R
\name{LOS_bus_interval}
\alias{LOS_bus_interval}
\title{Service level by region according to bus service interval(LOS_bus_interval)}
\usage{
LOS_bus_interval(city = NULL, bus_interval = NULL)
}
\arguments{
\item{city}{Choose one from: \code{'large'}, \code{'small'}}

\item{bus_interval}{bus operation interval(minutes)}
}
\description{
It follows <Table 13-5> in KHCM(2013), p.595.
}
\examples{
LOS_bus_interval('large', 23)
LOS_bus_interval('small', 92)
}
