% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_bus_seat.R
\name{LOS_bus_seat}
\alias{LOS_bus_seat}
\title{In-vehicle service level of seat-type bus(LOS_bus_seat)}
\usage{
LOS_bus_seat(n_psg = NULL, psg_by_seat = NULL, area_psg = NULL)
}
\arguments{
\item{n_psg}{Number of passenger}

\item{psg_by_seat}{Number of occupants per seat}

\item{area_psg}{Area occupied per passenger}
}
\description{
It follows <Table 13-3> in KHCM(2013), p.593.
It based on the vehicle area of 23.30㎡, standing area of 3.77㎡, 45 seats.
}
\examples{
LOS_bus_seat(n_psg = 30)
LOS_bus_seat(psg_by_seat = 0.98)
LOS_bus_seat(area_psg = 0.33)
}
