% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_bus_standing.R
\name{LOS_bus_standing}
\alias{LOS_bus_standing}
\title{In-vehicle service level of standing-type bus(LOS_standing_bus)}
\usage{
LOS_bus_standing(
  type = NULL,
  n_psg = NULL,
  psg_by_seat = NULL,
  area_psg = NULL
)
}
\arguments{
\item{type}{Choose one from: \code{'city'}, \code{'circular'}}

\item{n_psg}{Number of passenger}

\item{psg_by_seat}{Number of occupants per seat}

\item{area_psg}{Area occupied per passenger}
}
\description{
It follows <Table 13-4> in KHCM(2013), p.594.
    It based on following standard.
    * City Bus : vehicle area of 26.37㎡, standing area of 7.5㎡, 31 seats.
    * Circular Bus : vehicle area of 16.05㎡, standing area of 3.48㎡, 24 seats.
}
\examples{
LOS_bus_standing('city', 30)
LOS_bus_standing(type = 'circular', psg_by_seat = 1.334)
LOS_bus_standing(type = 'city', area_psg = 2.31)
}
