% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_bus_time.R
\name{LOS_bus_time}
\alias{LOS_bus_time}
\title{Service level according to bus operation time(LOS_bus_time)}
\usage{
LOS_bus_time(t = NULL)
}
\arguments{
\item{t}{operation time per day(hour)}
}
\description{
It follows <Table 13-6> in KHCM(2013), p.595.
}
\examples{
LOS_bus_time(19)
}
