% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_diamond_interchange.R
\name{LOS_diamond_interchange}
\alias{LOS_diamond_interchange}
\title{Level of Service(LOS) in Diamond-Interchange}
\usage{
LOS_diamond_interchange(control_delay = NULL)
}
\arguments{
\item{control_delay}{Control delay per Car(second)}
}
\description{
This function decides Level of Service(LOS). It follows <Table 6-8>
}
\examples{
LOS_diamond_interchange(33.4427)
}
\keyword{LOS}
\keyword{Level}
\keyword{Service}
\keyword{of}
