% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_freeway_basic.R
\name{LOS_freeway_basic}
\alias{LOS_freeway_basic}
\title{Level of Service(LOS) in Basic Freeway Section}
\usage{
LOS_freeway_basic(
  density = NULL,
  design_speed = NULL,
  volume = NULL,
  v_c_ratio = NULL
)
}
\arguments{
\item{density}{\emph{Numeric} The density of the road(pcpkmpl)}

\item{design_speed}{\emph{Categorical} Choose one from : \code{120}, \code{100}, \code{80}}

\item{volume}{\emph{Numeric} Traffic Volume(pcphpl)}

\item{v_c_ratio}{\emph{Numeric} It should be more than zero.}
}
\description{
This function decides Level of Service(LOS) in the basic freeway section.
}
\examples{
DDHV(AADT = 1500, region = 'city', road = 'general', lane = 2)
LOS_freeway_basic(density = 30)
LOS_freeway_basic(design_speed = 120, v_c_ratio = 0.5)
LOS_freeway_basic(design_speed = 80, volume = 1000)
}
\keyword{Density}
\keyword{LOS}
\keyword{Level}
\keyword{Service}
\keyword{V/C}
\keyword{of}
\keyword{ratio}
