% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_freeway_ramp_junction.R
\name{LOS_freeway_ramp_junction}
\alias{LOS_freeway_ramp_junction}
\title{Level of Service(LOS) in Freeway Ramp Junction}
\usage{
LOS_freeway_ramp_junction(V = NULL, capacity = NULL, density = NULL)
}
\arguments{
\item{V}{*Numeric* The traffic volume of target road.}

\item{capacity}{*Numeric* The capacity of target road,}

\item{density}{*Numeric* The density of the road(pcpkmpl)}
}
\description{
This function decides Level of Service(LOS) in the freeway ramp junction. It follows <Table 4-3> in KHCM(2013).
}
\examples{
LOS_freeway_ramp_junction(V = 3000, capacity = 4000, density = 23)
LOS_freeway_ramp_junction(4000, 3000, 24)
}
