% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_freeway_weaving.R
\name{LOS_freeway_weaving}
\alias{LOS_freeway_weaving}
\title{Level of Service(LOS) in Freeway Weaving Section}
\usage{
LOS_freeway_weaving(type = NULL, density = NULL)
}
\arguments{
\item{type}{*Categorical* Choose one from : \code{'link_link'}, \code{'main_link'}}

\item{density}{*Numeric* The density of the road(pcpkmpl)}
}
\description{
This function decides Level of Service(LOS) in the freeway weaving section.
}
\examples{
LOS_freeway_weaving(type = 'link_link', density = 29)
LOS_freeway_weaving('main_link', density = 28)
}
