% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_multilane_road.R
\name{LOS_multilane_road}
\alias{LOS_multilane_road}
\title{Level of Service(LOS) in Multi-lane Road}
\usage{
LOS_multilane_road(
  type = NULL,
  design_speed = NULL,
  v_c_ratio = NULL,
  free_speed = NULL,
  speed = NULL,
  g_c_ratio = NULL,
  service_volume = NULL
)
}
\arguments{
\item{type}{\emph{Categorical} Type of traffic flow in multi-lane road. Choose one from : \code{'interrupted_flow'}, \code{'uninterrupted_flow'}}

\item{design_speed}{\emph{Categorical} If \code{type == 'uninterrupted_flow'}, design_speed must be used.}

\item{v_c_ratio}{\emph{Numeric} V/C ratio. It is neccessary when \code{type = 'uninterrupted_flow'}}

\item{free_speed}{\emph{Numeric} If \code{type == 'interrupted_flow'}, It means free-speed(kph).}

\item{speed}{\emph{Numeric} It means speed(kph)}

\item{g_c_ratio}{\emph{Numeric} Average green time ratio. It is neccessary when \code{type == 'interrupted_flow'}}

\item{service_volume}{\emph{Numeric} Service Traffic Volume(pcphpl)}
}
\description{
This function decides Level of Service(LOS). It follows <Table 6-8>
}
\examples{
LOS_multilane_road(type = 'uninterrupted_flow', design_speed = 100, v_c_ratio = 0.20, service_volume = 500, speed = 100)
LOS_multilane_road(type = 'interrupted_flow', v_c_ratio = '0.74', free_speed = 70, speed = 59, g_c_ratio = 0.5, service_volume = 900)
}
\keyword{Density}
\keyword{LOS}
\keyword{Level}
\keyword{Service}
\keyword{V/C}
\keyword{of}
\keyword{ratio}
