% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_roundabout.R
\name{LOS_roundabout}
\alias{LOS_roundabout}
\title{Level of Service(LOS) in Roundabout(LOS_roundabout)}
\usage{
LOS_roundabout(avg_d = NULL, V_c_ratio = NULL)
}
\arguments{
\item{avg_d}{Average Delay(seconds/unit)}

\item{V_c_ratio}{Traffic Volume/Capacity}
}
\description{
This function follows <Table 11-1> in KHCM(2013) p.494
}
\examples{
LOS_roundabout(43.2, 0.3)
}
