% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_signalized_intersection.R
\name{LOS_signalized_intersection}
\alias{LOS_signalized_intersection}
\title{Level of Service(LOS) in Signalized Intersections.}
\usage{
LOS_signalized_intersection(control_delay = NULL)
}
\arguments{
\item{control_delay}{\emph{Numeric} (s) Delay due to the reduction or stop of the lane group due to signal control, The increase in travel time compared to travel time when there is no deceleration or stop.}
}
\description{
This function decides Level of Service(LOS). It follows <Table 8-2>
}
\examples{
LOS_signalized_intersection(control_delay = 44)
LOS_signalized_intersection(157)
}
\keyword{LOS}
\keyword{Level}
\keyword{Service}
\keyword{of}
