% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_two_way_stop.R
\name{LOS_two_way_stop}
\alias{LOS_two_way_stop}
\title{Level of Service(LOS) in two-way stop intersection(LOS_two_way_stop)}
\usage{
LOS_two_way_stop(avg_d = NULL)
}
\arguments{
\item{avg_d}{Average operating delay (sec/veh)}
}
\description{
This function follows <Table 10-2> in KHCM(2013) p.470
}
\examples{
LOS_two_way_stop(avg_d = 33.2)
LOS_two_way_stop(18.45)
}
