% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_uncontrolled_intersection.R
\name{LOS_uncontrolled_intersection}
\alias{LOS_uncontrolled_intersection}
\title{Level of Service(LOS) in uncontrolled intersection(LOS_uncontrolled_intersection)}
\usage{
LOS_uncontrolled_intersection(r = NULL, V = NULL)
}
\arguments{
\item{r}{main_road_traffic_ratio (%)}

\item{V}{Intersection total traffic (vph)}
}
\description{
This function follows <Table 10-3>, <Table 10-4>, <Formula 10-6> in KHCM(2013) p.473
}
