% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFi.R
\name{MSFi}
\alias{MSFi}
\title{Maximum Service Flow Rate(MSFi)}
\usage{
MSFi(design_speed = NULL, v_c_ratio_i = NULL)
}
\arguments{
\item{design_speed}{\emph{Numeric} Choose one from : \code{120}, \code{100}, \code{80}}

\item{v_c_ratio_i}{\emph{Numeric} V/c ratio under LOS i. It must be more than 0.}
}
\description{
This function calculates Maximum Service Flow Rate(MSFi) under basic conditions.
}
\examples{
MSFi(c_j = 1400, v_c_ratio_i = 0.2)
}
