% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_B.R
\name{mu_B}
\alias{mu_B}
\title{The speed at which the end of the queue moves to the end of the internal link (mu_B, m/sec)}
\usage{
mu_B(S_u_i = NULL, mu_u_i = NULL, l_Q = NULL)
}
\arguments{
\item{S_u_i}{Saturated traffic flow rate (= discharge traffic flow rate) of upstream intersection lane group i (car/sec)}

\item{mu_u_i}{Free passage speed of lane group i mitted from the upstream intersection (m/sec)}

\item{l_Q}{Average distance between vehicles in the queue (from the rear bumper of the vehicle in front to the rear bumper of the vehicle behind) (m)}
}
\description{
TIt follows <Formula 9-3> in KHCM(2013), p.429
}
