% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_F.R
\name{mu_F}
\alias{mu_F}
\title{The speed at which the end of the queue moves to the end of the internal link (mu_F, m/sec)}
\usage{
mu_F(S_d_i = NULL, mu_d_i = NULL, l_Q = NULL)
}
\arguments{
\item{S_d_i}{Saturated traffic flow rate (= discharge traffic flow rate) of downstream intersection lane group i (car/sec)}

\item{mu_d_i}{Free passage speed of lane group i mitted from the downstream intersection (m/sec)}

\item{l_Q}{Average distance between vehicles in the queue (from the rear bumper of the vehicle in front to the rear bumper of the vehicle behind) (m)}
}
\description{
TIt follows <Formula 9-5> in KHCM(2013), p.429
}
