% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_bus_stopping_area.R
\name{n_bus_stopping_area}
\alias{n_bus_stopping_area}
\title{Number of Bus Stop Stopping area(n_bus_stopping_area)}
\usage{
n_bus_stopping_area(l = NULL)
}
\arguments{
\item{l}{Length of bus stop(m)}
}
\description{
It follows KHCM(2013), p.597
}
\examples{
n_bus_stopping_area(12)
n_bus_stopping_area(11.999)
n_bus_stopping_area(25.142)
}
