% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_bus_using_efficiency.R
\name{N_bus_using_efficiency}
\alias{N_bus_using_efficiency}
\title{Utilization efficiency coefficient according to the stopping surface(N)}
\usage{
N_bus_using_efficiency(l = NULL)
}
\arguments{
\item{l}{length of the bus stop(m)}
}
\description{
It follows <Table 13-9> KHCM(2013), p.599
}
\examples{
N_bus_using_efficiency(43.22)
}
