% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_lanes_required.R
\name{number_of_lanes_required}
\alias{number_of_lanes_required}
\title{Number of lanes required(N)}
\usage{
number_of_lanes_required(PDDHV = NULL, SFi = NULL)
}
\arguments{
\item{PDDHV}{\emph{Numeric} Peak Directional Design Hourly Volume(PDDHV, vph). It must be positive.}

\item{SFi}{\emph{Numeric} #' Service Flow Rate(SFi) It must be positive.}
}
\description{
This function calculates the numver of lanes required(N) from PDDHV and SFi. It returns an integer.
}
\examples{
number_of_lanes_required(PDDHV = 14000, SFi = 3000)
number_of_lanes_required(3000, 500)
}
