% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_i_nonsignalize_intersection.R
\name{p_i_nonsignalize_intersection}
\alias{p_i_nonsignalize_intersection}
\title{Resistance coefficient against disturbance flow(p_i_nonsignalize_intersection)}
\usage{
p_i_nonsignalize_intersection(x = NULL)
}
\arguments{
\item{x}{Capacity ratio according to demand}
}
\description{
It follows <Formula 10-3> and <Figure 10-9> in KHCM(2013), p.469
}
