% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_L.R
\name{P_L}
\alias{P_L}
\title{The ratio of left turns in the group of practically dedicated left-turn lanes(P_L)}
\usage{
P_L(V_L = NULL, V_LF = NULL)
}
\description{
This function follows <Formula 8-21> in KHCM(2013)
}
\examples{
P_L(320, 832)
}
