% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_LT.R
\name{P_LT}
\alias{P_LT}
\title{Straight. Ratio of left turns in left-turning common lane groups(P_LT)}
\usage{
P_LT(lane_type = NULL, V_L = NULL, V_R = NULL, V_TH = NULL, V_RF = NULL)
}
\arguments{
\item{lane_type}{\emph{Categorical} Choose from one : \code{'public_left_turn_lane'}, \code{'integrated_lane'}}
}
\description{
This function follows <Formula 8-23>, <Formula 8-25> in KHCM(2013)
}
\examples{
P_LT(lane_type = 'public_left_turn_lane', V_L = 242, V_TH = 573, V_RF = 201)
}
