% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_R.R
\name{P_R}
\alias{P_R}
\title{The ratio of left turns in the group of practically dedicated right-turn lanes(P_R)}
\usage{
P_R(V_R = NULL, V_RF = NULL)
}
\description{
This function follows <Formula 8-22> in KHCM(2013)
}
\examples{
P_R(V_R = 321, V_RF = 1289)
}
