% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_RT.R
\name{P_RT}
\alias{P_RT}
\title{Straight. Ratio of left turns in right-turning common lane groups(P_RT)}
\usage{
P_RT(lane_type = NULL, V_L = NULL, V_R = NULL, V_TH = NULL, V_LF = NULL)
}
\arguments{
\item{lane_type}{\emph{Categorical} Choose from one : \code{'public_left_turn_lane'}, \code{'integrated_lane'}}
}
\description{
This function follows <Formula 8-24>, <Formula 8-25> in KHCM(2013)
}
\examples{
P_RT('public_left_turn_lane', 472, 382, 532, 222)
}
