% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_signalized_intersection.R
\name{P_signalized_intersection}
\alias{P_signalized_intersection}
\title{Number of possible unprotected left turns per gap by opposite traffic volume(P_signalized_intersection)}
\usage{
P_signalized_intersection(V_o = NULL)
}
\arguments{
\item{V_o}{*Numeric* Opposite traffic volume(vph)}
}
\description{
This function follows <Table 8-8> in KHCM(2013)
}
\examples{
P_signalized_intersection(V_o = 1792)
P_signalized_intersection(993)
}
