% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDDHV.R
\name{PDDHV}
\alias{PDDHV}
\title{Peak Directional Design Hourly Volume(PDDHV, vph)}
\usage{
PDDHV(AADT = NULL, region = NULL, road = NULL, lane = NULL, PHF = NULL)
}
\arguments{
\item{AADT}{*Numeric* Average Annual Daily Traffic(pc/day, vph)}

\item{region}{*Categorical* Choose one from : \code{'city'}, \code{'rural'}}

\item{road}{*Categorical* Choose one from : \code{'general'}, \code{'highway'}}

\item{lane}{*Numeric* The number of lane. It should \code{2} or \code{4} or more.}
}
\description{
This function calculates the Peak Directional Design Hourly Volume(PDDHV, vph) for a given road and traffic condition.
}
\examples{
PDDHV(AADT = 1000, region = 'city', road = 'general', lane = 6, PHF = 0.98)
PDDHV(1500, 'rural', 'highway', 8, 0.95)
}
\keyword{AADT}
\keyword{DDHV}
\keyword{DHV}
\keyword{Design}
\keyword{Directional}
\keyword{Factor}
\keyword{Hourly}
\keyword{PDDHF}
\keyword{PHF}
