% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF_fixed.R
\name{PF_fixed}
\alias{PF_fixed}
\title{Fixed time signal interlocking factor (PF_fixed)}
\usage{
PF_fixed(TVO = NULL, g_c_ratio = NULL)
}
\description{
This function follows <Table 8-17>, <12-8> in KHCM(2013) p.541
}
\examples{
PF_fixed(TVO = 0.4, g_c_ratio = 0.42)
PF_fixed(0.5, 0.2223)
}
