% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF_sensitive.R
\name{PF_sensitive}
\alias{PF_sensitive}
\title{Interlocking coefficient of response signal(PF_sensitive)}
\usage{
PF_sensitive(
  type = NULL,
  direction = NULL,
  turn = NULL,
  arrival_type = NULL,
  v_c_ratio = NULL
)
}
\arguments{
\item{type}{Signal Type. Choose one from: \code{'sensitive'}, \code{'semi-sensitive'}}

\item{direction}{\code{'main'}, \code{'sub'}}

\item{turn}{\code{'straight'}, \code{'right'}, \code{'simultaneous_left'}, \code{'dedicated_left'}}

\item{arrival_type}{Choose one from: \code{1}, \code{2}, \code{3}, \code{4}, \code{5}}
}
\description{
It follows <Table 12-9> in KHCM(2013) p.542
}
\examples{
PF_sensitive('semi-sensitive', 'main', 'right', 3, 0.23)
}
