% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHF_2lane_road.R
\name{PHF_2lane_road}
\alias{PHF_2lane_road}
\title{Peak Hour Factor in 2-lane Road (PHF_2lane_road)}
\usage{
PHF_2lane_road(V = NULL)
}
\arguments{
\item{V}{*Numeric* Traffic Volume(pcph/bidirectional)}
}
\description{
This function calculates the peak hour factor(PHF) in 2-lane-road in given traffic volume.
    It is desirable to measure the PHF in the field, and this table is a general value,
    so it should be used in a limited way.
}
\examples{
PHF_2lane_road(V = 3020)
PHF_2lane_road(2000)
}
