% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_coefficient.R
\name{r_coefficient}
\alias{r_coefficient}
\title{Compensation factor for calculating the stopping surface capacity according to the bus waiting ratio (r_coefficient)}
\usage{
r_coefficient(t_R = NULL)
}
\arguments{
\item{t_R}{Bus waiting rate(\%)}
}
\description{
It follows <Table 13-9> KHCM(2013), p.599
}
\examples{
r_coefficient(13.29)
}
