% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadside_friction_arterial.R
\name{roadside_friction_arterial}
\alias{roadside_friction_arterial}
\title{Degree of roadside friction on arterial roads(roadside_friction_arterial)}
\usage{
roadside_friction_arterial(
  free_speed = NULL,
  bus_stop = NULL,
  entrance_exit = NULL
)
}
\arguments{
\item{free_speed}{Choose one from : \code{80}, \code{70}, \code{60}}

\item{bus_stop}{Number of bus stop(ea/km)}

\item{entrance_exit}{Number of entrance and exit(ea/km)}
}
\description{
It follows <Table 12-6> in KHCM(2013) p.536
}
\examples{
roadside_friction_arterial(80, 2, 3)
}
