% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_avg_weaving.R
\name{S_avg_weaving}
\alias{S_avg_weaving}
\title{Average Speed in Freeway Weaving Section(S_avg_weaving, kph)}
\usage{
S_avg_weaving(V = NULL, V_w = NULL, V_nw = NULL, S_w = NULL, S_nw = NULL)
}
\arguments{
\item{V}{\emph{Numeric} Total traffic volume of the weaving section(pcph)}

\item{V_w}{\emph{Numeric} Weaving traffic volume(pcph)}

\item{V_nw}{\emph{Numeric} Non-weaving traffic volume(pcph)}

\item{S_w}{\emph{Numeric} Average Speed of Weaving Traffic Flow in Freeway Weaving Section(S_w, kph)}

\item{S_nw}{\emph{Numeric} Average Speed of Non-Weaving Traffic Flow in Freeway Weaving Section(S_nw, kph)}
}
\description{
This function calculates the average speed in freeway weaving section.(kph)
}
\examples{
S_avg_weaving(V = 1000, V_w = 300, V_nw = 700, S_w = 80, S_nw = 75)
S_avg_weaving(1100, 350, 750, 85, 80)
}
