% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_bi.R
\name{S_bi}
\alias{S_bi}
\title{Saturated traffic flow rate (vphg) of lane group i in the exclusive median bus lane(S_bi, vphg)}
\usage{
S_bi(N_i = NULL, f_BLT = NULL, f_g = NULL, f_ub = NULL)
}
\arguments{
\item{N_i}{the number of lanes in the i lane group}

\item{f_BLT}{Left turn correction factor on exclusive median bus lane (1.0 in case of going straight)}

\item{f_g}{Approach slope correction factor}

\item{f_ub}{Upstream stop influence correction factor}
}
\description{
It follows <Formula 8-58> in KHCM(2013)
}
