% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_i.R
\name{S_i}
\alias{S_i}
\title{Saturated traffic flow rate of lane group i (S_ivphg)}
\usage{
S_i(N_i = NULL, f_LRT = NULL, f_w3 = NULL, f_g = NULL, f_hw3 = NULL, f_IW = 1)
}
\arguments{
\item{N_i}{number of lanes in the i lane group}

\item{f_LRT}{f_LT or f_RT}

\item{f_w3}{Lane width correction factor}

\item{f_g}{Approach slope correction factor}

\item{f_IW}{weather correction factor}

\item{f_hv3}{Heavy vehicle factor}
}
\description{
It follows <Formula 8-38>, <Formula 8-68> in KHCM(2013)
}
