% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_i_backtick.R
\name{S_i_backtick}
\alias{S_i_backtick}
\title{Saturated traffic flow rate of i-lane group reflecting special circumstances (construction section, weather). vphg, S_i_backtick}
\usage{
S_i_backtick(S_i = NULL, f_WZ = NULL, f_IW = NULL)
}
\arguments{
\item{S_i}{i Saturated traffic flow rate(vphg)}

\item{f_WZ}{Construction impact correction factor}

\item{f_IW}{weather correction factor}
}
\description{
It follows <Formula 8-38> in KHCM(2013)
}
