% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_i_WZ.R
\name{S_i_WZ}
\alias{S_i_WZ}
\title{Correction of saturated traffic flow rate due to road occupation construction near intersections(S_i_WZ)}
\usage{
S_i_WZ(
  N_i = NULL,
  f_LRT = NULL,
  f_w3 = NULL,
  f_g = NULL,
  f_hw3 = NULL,
  f_WZ = NULL
)
}
\arguments{
\item{N_i}{number of lanes in the i lane group}

\item{f_w3}{Lane width correction factor}

\item{f_g}{Approach slope correction factor}

\item{f_hv3}{Heavy vehicle factor}
}
\description{
It follows <Formula 8-66> in KHCM(2013)
}
