% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_nw.R
\name{S_nw}
\alias{S_nw}
\title{Average Speed of Non-Weaving Traffic Flow in Freeway Weaving Section(S_nw, kph)}
\usage{
S_nw(S_D = NULL, V = NULL, V_w = NULL, N = NULL, L = NULL)
}
\arguments{
\item{S_D}{\emph{Numeric} Design speed of the main lane(kph)}

\item{V}{\emph{Numeric} Total traffic volume (pcph) of the weaving section}

\item{V_w}{\emph{Numeric} Weaving traffic volume(pcph)}

\item{N}{\emph{Numeric} Total number of lanes in the weaving section}

\item{L}{\emph{Numeric} Length of the weaving section(m)}
}
\description{
This function calculates the average speed of non-weaving traffic flow in freeway weaving section.(kph)
}
\examples{
S_nw(S_D = 100, V = 1000, V_w = 300, N = 1, L = 250)
S_nw(120, 1200, 210, 1, 300)
}
