% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_w.R
\name{S_w}
\alias{S_w}
\title{Average Speed of Weaving Traffic Flow in Freeway Weaving Section(S_w, kph)}
\usage{
S_w(S_D = NULL, V = NULL, V_w = NULL, N = NULL, L = NULL)
}
\arguments{
\item{S_D}{*Numeric* Design speed of the main lane(kph)}

\item{V}{*Numeric* Total traffic volume (pcph) of the weaving section}

\item{V_w}{*Numeric* Weaving traffic volume(pcph)}

\item{N}{*Numeric* Total number of lanes in the weaving section}

\item{L}{*Numeric* Length of the weaving section(m)}

\item{N_main}{*Numeric* A Number of main lane in the weaving section}
}
\description{
This function calculates the average speed of weaving traffic flow in freeway weaving section.(kph)
}
\examples{
S_w(S_D = 100, V = 1000, V_w = 300, N = 4, N_main = 3 L = 250)
S_w(120, 1200, 210, 3, 2, 300)
}
