% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFi.R, R/v_c_ratio.R
\name{SFi}
\alias{SFi}
\title{Service Flow Rate(SFi)}
\usage{
SFi(c_j = NULL, v_c_ratio_i = NULL, N = NULL, f_w = NULL, f_hv = NULL)

SFi(c_j = NULL, v_c_ratio_i = NULL, N = NULL, f_w = NULL, f_hv = NULL)
}
\arguments{
\item{v_c_ratio_i}{\emph{Numeric} V/c ratio under LOS i.}

\item{N}{\emph{Numeric} The number of freeway lane(one-way lane). It must be 2 or more.}

\item{f_w}{\emph{Numeric} The One Side Lane Width and Lateral Clearance Factor(f_w). It could be calculated by f_w()}

\item{f_hv}{\emph{Numeric} Heavy Vehicle Factors(f_hv). See f_hv()}

\item{design_speed}{\emph{Numeric} Choose one from : \code{120}, \code{100}, \code{80}}

\item{Vp}{\emph{Numeric} Traffic Volume Converted to Peak Hour(Vp, vph)}

\item{c}{\emph{Numeric} Capacity}
}
\description{
This function calculates Service Flow Rate(MSFi) under basic conditions.

The Volume-to-Capacity ratio (Vp/C) measures the level of congestion on a roadway by dividing the volume (VPD) of traffic (existing or future) by the capacity of the roadway.
}
\examples{
SFi(c_j = 1400, v_c_ratio_i = 0.2, N = 6, f_w = 0.5, f_hv = 0.8)
SFi(c_j = 1400, v_c_ratio_i = 0.2, N = 6, f_w = 0.5, f_hv = 0.8)
}
