% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_b.R
\name{T_b}
\alias{T_b}
\title{Saturation head time increment according to one bus stop (T_b, sec)}
\usage{
T_b(bus_stop_loc = NULL, bus_passenger_size = NULL)
}
\arguments{
\item{bus_stop_loc}{*Categorical* Choose one from : \code{'driving_lane'}, \code{'separated'}}

\item{bus_passenger_size}{*Categorical* Choose one from: \code{'small'}, \code{'middle'}, \code{'large'}}
}
\description{
This function follows <Table 8-12> in KHCM(2013)
}
\examples{
T_b('driving_lane', 'middle')
T_b('separated')
}
