% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_bus.R
\name{T_bus}
\alias{T_bus}
\title{Central bus lane analysis section travel time (T_bus, seconds)}
\usage{
T_bus(t_p_km_arterial_central_bus_lane = NULL, l = NULL, d = NUL)
}
\arguments{
\item{t_p_km_arterial_central_bus_lane}{see function. Total travel time (sec/km) per km of some sections of exclusive median bus lanes or general lanes}

\item{l}{Length of analysis section (km)}

\item{d}{Total access delay (s) at all intersections within the analysis target range as a part of the exclusive median bus lane or general lane}
}
\description{
It follows <Formula 12-11> in KHCM(2013) p.552.
}
