% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_c_x.R
\name{t_c_x}
\alias{t_c_x}
\title{Critical interval of two-way stop intersection(t_c_x)}
\usage{
t_c_x(form = NULL, direction = NULL)
}
\arguments{
\item{form}{Choose one from: \code{'1x1'}, \code{'2x1'}}

\item{direction}{Choose one from: \code{'main_left'}, \code{'sub_left'}, \code{'sub_straight'}, \code{'sub_right'}}
}
\description{
TIt follows <Table 10-1> in KHCM(2013), p.466
}
\examples{
t_c_x(form = '2x1', direction = 'sub_straight')
t_c_x('1x1', 'main_left')
}
