% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_D.R
\name{t_D}
\alias{t_D}
\title{bus stop time. Based on door opening time and passenger boarding and disembarking time(t_D, seconds)}
\usage{
t_D(type = NULL, standing_psg = NULL, pay = NULL)
}
\arguments{
\item{type}{Bus type. Choose one from: \code{'city'}, \code{'seat'}}

\item{standing_psg}{Choose one from: \code{'yes'}, \code{'no'}}

\item{pay}{Choose one from: \code{'card'}, \code{'cash_change'}, \code{'cash_no_change'}}
}
\description{
It follows <Table 13-7> in KHCM(2013), p.597.
}
\examples{
t_D(type = 'city', standing_psg = 'yes', pay = 'card')
t_D('seat', 'no', pay = 'cash_no_change')
}
