% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_other.R
\name{T_other}
\alias{T_other}
\title{General lane analysis section travel time (T_bus, seconds)}
\usage{
T_other(t_p_km_arterial = NULL, l = NULL, d = NUL)
}
\arguments{
\item{t_p_km_arterial}{see function. Total travel time (sec/km) per km of some sections of general lanes}

\item{l}{Length of analysis section (km)}

\item{d}{Total access delay (s) at all intersections within the analysis target range as a part of the general lane}
}
\description{
It follows <Formula 12-11> in KHCM(2013) p.552.
}
