% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_p_km_arterial.R
\name{t_p_km_arterial}
\alias{t_p_km_arterial}
\title{Section travel time per km(t_p_km_arterial, seconds/km)}
\usage{
t_p_km_arterial(free_speed = NULL, roadside_friction_arterial = NULL, L = NULL)
}
\arguments{
\item{free_speed}{Choose one from : \code{80}, \code{70}, \code{60}}

\item{roadside_friction_arterial}{Choose one from: \code{'big'}, \code{'small'}. See roadside_friction_arterial()}

\item{L}{section length(km)}
}
\description{
It follows <Table 12-5> in KHCM(2013) p.536
}
\examples{
t_p_km_arterial(free_speed = 80, roadside_friction_arterial = 'big', L = 0.83)
t_p_km_arterial(60, 'small', 0.213)
}
