% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_p_km_arterial_central_bus_lane.R
\name{t_p_km_arterial_central_bus_lane}
\alias{t_p_km_arterial_central_bus_lane}
\title{Section travel time per km in Arterial Road with Central Bus lane(t_p_km_arterial_central_bus_lane, seconds/km)}
\usage{
t_p_km_arterial_central_bus_lane(
  L = NULL,
  overtaking_lane = NULL,
  N_bus_stop = NULL
)
}
\arguments{
\item{L}{Interval between analysis sections (km)}

\item{overtaking_lane}{Choose one from: \code{'yes'}, \code{'no'}}

\item{N_bus_stop}{Number of bus stop. \code{0}, \code{1}, \code{2}}
}
\description{
It follows <Table 12-11> in KHCM(2013) p.550
}
\examples{
t_p_km_arterial_central_bus_lane(0.32, 'yes', 1)
}
