% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_0.R
\name{tau_0}
\alias{tau_0}
\title{Time (in seconds) from the start until the first vehicle emitted from the upstream intersection arrives at the end of the queue(tau_0)}
\usage{
tau_0(D = NULL, Q_L = NULL, mu_u_i = NULL)
}
\arguments{
\item{D}{Length of internal link between intersections (m)}

\item{Q_L}{Initial queue length (m) by appearance of internal link between intersections. Q_V * l_Q}

\item{mu_u_i}{Free passage speed of lane group i mitted from the upstream intersection (m/sec)}
}
\description{
TIt follows <Formula 9-2> in KHCM(2013), p.429
}
