% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_1.R
\name{tau_1}
\alias{tau_1}
\title{Time (in seconds) from t_1  when the queue arrives at the end of the internal link(tau_1)}
\usage{
tau_1(D = NULL, Q_L = NULL, mu_B = NULL)
}
\arguments{
\item{D}{Length of internal link between intersections (m)}

\item{Q_L}{Initial queue length (m) by appearance of internal link between intersections. Q_V * l_Q}

\item{mu_B}{The speed at which the end of the queue moves to the end of the internal link (m/sec)}
}
\description{
TIt follows <Formula 9-3> in KHCM(2013), p.429
}
