% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVO.R
\name{TVO}
\alias{TVO}
\title{Offset convenience rate(TVO)}
\usage{
TVO(T_c = NULL, offset = NULL, C = NULL)
}
\arguments{
\item{T_c}{In the section from the stop line of the upstream intersection to the stop line of the analysis intersection,Time (seconds) obtained from the speed and link length of the section not affected by acceleration, deceleration, and stop}

\item{offset}{The difference (in seconds) between the upstream intersection and the analysis intersection in the continuous direction green signal start time. Use less value than period}

\item{C}{Common cycle required for arterial linkage (seconds)}
}
\description{
It follows <Formula 8-51>, <Formula 8-52>, <Formula 8-53>, <12-8> in KHCM(2013) p.540
}
