% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_d3exist.R
\name{type_d3exist}
\alias{type_d3exist}
\title{3 types of extra lag d3 present(type_d3exist)}
\usage{
type_d3exist(Q_b = NULL, X = NULL, C = NULL, T = NULL)
}
\description{
It follows <Formula 8-44>, <Formula 8-45>, <Formula 8-46> in KHCM(2013)
}
