% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_c_i_roundabout.R
\name{V_c_i_roundabout}
\alias{V_c_i_roundabout}
\title{Conflict traffic volume on approach i (V_c_i_roundabout, pcph)}
\usage{
V_c_i_roundabout(V_ip1 = NULL, V_ip2 = NULL, V_ip3 = NULL, V_ip4 = NULL)
}
\arguments{
\item{V_ip1}{Traffic from Approach i+1 to Approach i+1}

\item{V_ip2}{Traffic volume moving from approach i+4 to each approach i+1, i+2}

\item{V_ip3}{Traffic volume moving from approach i+4 to each approach i+1, i+2, i+3}

\item{V_ip4}{Traffic volume moving from approach i+4 to each approach i+1, i+2, i+3, i+4}
}
\description{
It follows <Formula 11-5> in KHCM(2013) p.500
}
