% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_c_NB.R
\name{V_c_NB}
\alias{V_c_NB}
\title{Roundabout ramp capacity(c_roundabout, pcph)}
\usage{
V_c_NB(V_EBL = NULL, V_EBT = NULL, V_SBL = NULL, V_U_NB = NULL)
}
\arguments{
\item{V_EBL}{East-turn left-entry traffic (vph)}

\item{V_EBT}{Traffic volume going straight in the east direction (vph)}

\item{V_SBL}{South-facing left turn traffic volume (vph)}

\item{V_U_NB}{Sum of U-turn traffic on each approach except for northbound U-turn traffic (vph)}
}
\description{
It follows <Formula 11-2> in KHCM(2013) p.493
}
