% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_ci.R
\name{V_ci}
\alias{V_ci}
\title{Conflict Traffic Flow(V_ci)}
\usage{
V_ci(
  type = NULL,
  main_right_turn_lane = NULL,
  sub_road_wide = NULL,
  main_multi_load = NULL,
  V_r = NULL,
  V_t = NULL,
  V_ra = NULL,
  V_ta = NULL,
  V_la = NULL,
  V_rb = NULL,
  V_tb = NULL,
  V_lb = NULL,
  V_o = NULL,
  V_or = NULL
)
}
\arguments{
\item{type}{Choose one from : \code{'right_sub'}, \code{'left_main'}, \code{'straight_sub'}, \code{'left_sub'}}

\item{main_right_turn_lane}{\code{'yes'}, \code{'no'}}

\item{sub_road_wide}{\code{'yes'}, \code{'no'}. Where the radius of the right turn on the sub-road is wide, or where this flow is controlled by stopping / yielding}

\item{main_multi_load}{\code{'yes'}, \code{'no'}}
}
\description{
It follows <Figure 10-5> in KHCM(2013), p.465
}
