% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_e_i_roundabout.R
\name{V_e_i_roundabout}
\alias{V_e_i_roundabout}
\title{i-direction entry traffic in roundabout(V_e_i_roundabout, pcph)}
\usage{
V_e_i_roundabout(
  right_turn_lane = NULL,
  V_i_L = NULL,
  V_i_T = NULL,
  V_i_R = NULL,
  V_i_U = NULL
)
}
\arguments{
\item{right_turn_lane}{Choose one from: \code{'yes'}, \code{'no'}}

\item{V_i_L}{i-direction left turn traffic volume(pcph)}

\item{V_i_T}{i-direction straight traffic volume(pcph)}

\item{V_i_R}{i-direction right turn traffic volume(pcph)}

\item{V_i_U}{i-direction U-turn traffic volume(pcph)}
}
\description{
It follows <Formula 11-5> in KHCM(2013) p.500
}
