% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_i_pce.R
\name{V_i_pce}
\alias{V_i_pce}
\title{i Car traffic per hour on the ramp in roundabout(V_i_pce, pcph)}
\usage{
V_i_pce(lane = NULL, hv_ratio = NULL, P_T = NULL)
}
\arguments{
\item{lane}{Choose one from L \code{1}, \code{2}}

\item{hv_ratio}{Heavy Vehicle Ratio(%)}

\item{P_T}{Heavy Vehicle Mixing Ratio}

\item{V_p_i}{i ramp traffic volume (vph)}
}
\description{
It follows <Formula 11-4> in KHCM(2013) p.500
}
