% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_i_per_c_pi.R
\name{V_i_per_c_pi}
\alias{V_i_per_c_pi}
\title{Capacity ratio according to demand(V_i_per_c_pi)}
\usage{
V_i_per_c_pi(V_i = NULL, c_pi = NULL)
}
\arguments{
\item{V_i}{Traffic Volume of traffic flow i}

\item{c_pi}{Potential capacity for moving flow i}
}
\description{
This function follows <Formula 10-2> in KHCM(2013) p.468
}
