% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_LF.R
\name{V_LF}
\alias{V_LF}
\title{Number of vehicles going straight ahead of the first left turn in the public left turn lane(V_LF)}
\usage{
V_LF(V_L = NULL, V_TH = NULL, C = NULL, N = NULL)
}
\arguments{
\item{V_L}{\emph{Numeric} Left Turn Traffic Volume(vph)}

\item{V_TH}{\emph{Numeric} Straight-through traffic (vph)}

\item{C}{\emph{Numeric} Signal cycle(seconds)}

\item{N}{\emph{Categorical} Total number of access lanes (excluding dedicated left-turn lanes). \code{4}, \code{5}, \code{6}}
}
\description{
This function follows <Formula 8-13>, <Formula 8-14> in KHCM(2013)
}
\examples{
V_LF(423, 923, 76, 4)
}
