% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_RF.R
\name{V_RF}
\alias{V_RF}
\title{Number of vehicles going straight ahead of the first right turn in the public right turn lane(V_RF)}
\usage{
V_RF(V_R = NULL, V_TH = NULL, C = NULL, N = NULL)
}
\arguments{
\item{V_R}{\emph{Numeric} Right Turn Traffic Volume(vph)}

\item{V_TH}{\emph{Numeric} Straight-through traffic (vph)}

\item{C}{\emph{Numeric} Signal cycle(seconds)}

\item{N}{\emph{Categorical} Total number of access lanes (excluding dedicated left-turn lanes). \code{1}, \code{2}, \code{3}, \code{4}, \code{5}, \code{6}}
}
\description{
This function follows <Formula 8-15>, <Formula 8-16> in KHCM(2013)
}
\examples{
V_RF(432, 1293, 82, 4)
}
