% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_STL.R
\name{V_STL}
\alias{V_STL}
\title{Straight-through traffic (vph) using public left-turn lanes(V_STL)}
\usage{
V_STL(V_R = NULL, V_L = NULL, V_TH = NULL, E_R = NULL, E_L = NULL, N = NULL)
}
\arguments{
\item{V_R}{*Numeric* Right Turn Traffic Volume(vph)}

\item{V_L}{*Numeric* Left Turn Traffic Volume(vph)}

\item{V_TH}{*Numeric* Straight-through traffic (vph)}

\item{E_R}{*Numeric* Forward conversion factor for right turn}

\item{E_L}{*Numeric* Forward conversion factor for left turn}

\item{N}{*Categorical* Total number of access lanes (excluding dedicated left-turn lanes). \code{1}, \code{2}, \code{3}, \code{4}, \code{5}, \code{6}}
}
\description{
This function follows <Formula 8-17>, <Formula 8-18> in KHCM(2013)
}
\examples{
V_STL(V_R = 323, V_L = 291, V_TH = 999, E_R = 1.2, E_L = 1.09, N = 4)
}
