% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_STR.R
\name{V_STR}
\alias{V_STR}
\title{Straight-through traffic (vph) using public right-turn lanes(V_STR)}
\usage{
V_STR(V_R = NULL, V_L = NULL, V_TH = NULL, E_R = NULL, E_L = NULL, N = NULL)
}
\arguments{
\item{V_R}{*Numeric* Right Turn Traffic Volume(vph)}

\item{V_L}{*Numeric* Left Turn Traffic Volume(vph)}

\item{V_TH}{*Numeric* Straight-through traffic (vph)}

\item{E_R}{*Numeric* Forward conversion factor for right turn}

\item{E_L}{*Numeric* Forward conversion factor for left turn}

\item{N}{*Categorical* Total number of access lanes (excluding dedicated left-turn lanes). \code{1}, \code{2}, \code{3}, \code{4}, \code{5}, \code{6}}
}
\description{
This function follows <Formula 8-19>, <Formula 8-20> in KHCM(2013)
}
\examples{
V_STR(92, 103, 763, 1.3, 2.123, 5)
}
