% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vp.R
\name{Vp}
\alias{Vp}
\title{Traffic Volume Converted to Peak Hour(Vp)}
\usage{
Vp(V = NULL, PHF = NULL)
}
\arguments{
\item{V}{\emph{Numeric} Current or future traffic volume. It must be positive.}

\item{PHF}{\emph{Numeric} Peak Hour Factor(PHF). It is a value between 0 and 1.}
}
\description{
This function calculates traffic volume converted to peak hour(vph) by using PHF(peak hour factor).
}
\examples{
Vp(1000, 0.97)
Vp(V = 1300, PHF = 0.93)
}
