% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vp_i_roundabout.R
\name{Vp_i_roundabout}
\alias{Vp_i_roundabout}
\title{Traffic Volume Converted to Peak Hour in Roundabout(Vp_i_roundabout)}
\usage{
Vp_i_roundabout(V_i = NULL, PHF = NULL)
}
\arguments{
\item{V_i}{\emph{Numeric} i ramp traffic (vph)}

\item{PHF}{\emph{Numeric} Peak Hour Factor(PHF). It is a value between 0 and 1.}
}
\description{
This function calculates traffic volume converted to peak hour(vph) by using PHF(peak hour factor).
}
\examples{
Vp(1000, 0.97)
Vp(V = 1300, PHF = 0.93)
}
