% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/W_nw.R
\name{W_nw}
\alias{W_nw}
\title{Weaving Strength Factor According to Non-Weaving Traffic Flow(W_nw)}
\usage{
W_nw(V = NULL, V_w = NULL, N = NULL, L = NULL)
}
\arguments{
\item{V}{*Numeric* Total traffic volume (pcph) of the weaving section}

\item{V_w}{*Numeric* Weaving traffic volume(pcph)}

\item{N}{*Numeric* Total number of lanes in the weaving section}

\item{L}{*Numeric* Length of the weaving section(m)}
}
\description{
This function calculates the weaving strength factor according to non-weaving traffic flow(W_nw)
}
\examples{
W_nw(V = 1000, V_w = 300, N = 1, L = 100)
W_nw(1200, 100, 2, 200)
}
