% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_c.R
\name{X_c}
\alias{X_c}
\title{Critical V/c ratio(X_c)}
\usage{
X_c(C = NULL, L = NULL, y_i = NULL)
}
\arguments{
\item{C}{period(sec)}

\item{L}{Total loss time per cycle (sec)}

\item{y_i}{Traffic volume ratio for critical lane groups in each prefecture}
}
\description{
It follows <Formula 8-42> in KHCM(2013)
}
