% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_i.R
\name{X_i}
\alias{X_i}
\title{Saturation of lane group i(X_i)}
\usage{
X_i(S_i = NULL, g_i = NULL, C = NULL, V_i = NULL)
}
\arguments{
\item{S_i}{Saturated traffic flow rate (vph) for the i lane group}

\item{g_i}{Effective green time for i lane group (seconds)}

\item{C}{signal period(seconds)}

\item{V_i}{i traffic volume in lane group (vph)}
}
\description{
It follows <Formula 8-41> in KHCM(2013)
}
