% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_i_diamond_interchange.R
\name{X_i_diamond_interchange}
\alias{X_i_diamond_interchange}
\title{Saturation of the i lane group(X_i_diamond_interchange)}
\usage{
X_i_diamond_interchange(V_i = NULL, S_i = NULL, g_i_backtick = NULL, C = NULL)
}
\arguments{
\item{V_i}{Traffic volume of i lane group(vph)}

\item{S_i}{Saturated traffic flow rate (vph) for i lane group}

\item{g_i_backtick}{Effective green time (seconds) for the i lane group reflecting the additional green loss time}

\item{C}{period(seconds)}
}
\description{
It follows <Formula 9-10> in KHCM(2013) p.434
}
